% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_dataset_container}
\alias{add_dataset_container}
\title{Add organization to dataset}
\usage{
add_dataset_container(dataset, container_name, configuration = NULL)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{container_name}{charater, A valid RIDL container name, you can use
\code{list_containers()} to have the list of all containers}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A RIDLDataset
}
\description{
Add organization to dataset
}
\examples{
\dontrun{
 ds <- create_dataset(list(name = "cool-dataset"))
 add_dataset_container(ds, "zimbabwe-shelter-nfi")
}
}
