% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_dataset_resource}
\alias{add_dataset_resource}
\title{Add a Resource to a dataset}
\usage{
add_dataset_resource(
  dataset,
  resource,
  ignore_dataset_id = FALSE,
  configuration = NULL
)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{resource}{RIDLResource, the resource}

\item{ignore_dataset_id}{logical, ignore the dataset id}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A RIDLDataset
}
\description{
Add a Resource to a dataset
}
