% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{add_resource.RIDLDataset}
\alias{add_resource.RIDLDataset}
\alias{add_resource}
\alias{add_resource.default}
\title{Add a Resource to a dataset}
\usage{
\method{add_resource}{RIDLDataset}(
  dataset,
  resource,
  ignore_dataset_id = FALSE,
  configuration = NULL
)

add_resource(dataset, resource, ignore_dataset_id, configuration)

\method{add_resource}{default}(dataset, resource, ignore_dataset_id, configuration)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{resource}{RIDLResource, the resource}

\item{ignore_dataset_id}{logical, ignore the dataset id}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A RIDLDataset
}
\description{
Add a Resource to a dataset
}
