% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{Container}
\alias{Container}
\title{RIDL Container}
\description{
RIDL Container

RIDL Container
}
\details{
RIDL Container
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{Container}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for the Container fields element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Container$new()}}
\item \href{#method-get_datasets}{\code{Container$get_datasets()}}
\item \href{#method-browse}{\code{Container$browse()}}
\item \href{#method-as_list}{\code{Container$as_list()}}
\item \href{#method-print}{\code{Container$print()}}
\item \href{#method-clone}{\code{Container$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_configuration">}\href{../../ridl/html/RIDLObject.html#method-get_configuration}{\code{ridl::RIDLObject$get_configuration()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Container object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Container object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_datasets"></a>}}
\if{latex}{\out{\hypertarget{method-get_datasets}{}}}
\subsection{Method \code{get_datasets()}}{
Get the list of datasets within the container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of Dataset objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the Container page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with container field element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Container$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
