% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a RIDL dataset from list}
\usage{
create_dataset(initial_data, configuration = NULL)
}
\arguments{
\item{initial_data}{List, list of data}

\item{configuration}{RIDLConfig, RIDL configuration used}
}
\value{
Dataset the dataset
}
\description{
Create a RIDL dataset from list with required fields
}
\examples{
\dontrun{

 dsdata <- list(name = "hum-dataset",
                title = "Humanitarian dataset")
 res <- create_dataset(dsdata)
 res
}
}
