% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{create_resource}
\alias{create_resource}
\title{Create a RIDL resource from list}
\usage{
create_resource(initial_data, configuration = NULL)
}
\arguments{
\item{initial_data}{List, list of data}

\item{configuration}{RIDLConfig, the configuration used}
}
\value{
Resource the resource
}
\description{
Create a RIDL resource from list with required fields
}
\examples{
\dontrun{

 rsdata <- list(name = "hum-resource",
                title = "Humanitarian resource")
 res <- create_resource(rsdata)
 res
}
}
