% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{create_ridl_config}
\alias{create_ridl_config}
\title{Create a RIDL configuration object}
\usage{
create_ridl_config(ridl_key = NULL, user_agent = NULL)
}
\arguments{
\item{ridl_key}{Character for the CKAN API key, it is required to push data into RIDL}

\item{user_agent}{a character value, A user agent
you can push metdata and data to RIDL}
}
\value{
An RIDL Configuration object
}
\description{
Create and RIDL configuration object
}
