% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{Dataset}
\alias{Dataset}
\title{RIDL Dataset}
\description{
RIDL Dataset

RIDL Dataset
}
\details{
Dataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{Dataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resources}}{list of Resource object within the dataset}

\item{\code{data}}{placeholder for Dataset field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Dataset$new()}}
\item \href{#method-get_dataset_resource}{\code{Dataset$get_dataset_resource()}}
\item \href{#method-list_dataset_resources}{\code{Dataset$list_dataset_resources()}}
\item \href{#method-number_of_resources}{\code{Dataset$number_of_resources()}}
\item \href{#method-delete_dataset_resource}{\code{Dataset$delete_dataset_resource()}}
\item \href{#method-delete_all_dataset_resources}{\code{Dataset$delete_all_dataset_resources()}}
\item \href{#method-browse}{\code{Dataset$browse()}}
\item \href{#method-get_configuration}{\code{Dataset$get_configuration()}}
\item \href{#method-get_dataset_date_range}{\code{Dataset$get_dataset_date_range()}}
\item \href{#method-get_maintainer}{\code{Dataset$get_maintainer()}}
\item \href{#method-get_container}{\code{Dataset$get_container()}}
\item \href{#method-as_list}{\code{Dataset$as_list()}}
\item \href{#method-print}{\code{Dataset$print()}}
\item \href{#method-clone}{\code{Dataset$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Dataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_resource"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset_resource}{}}}
\subsection{Method \code{get_dataset_resource()}}{
Get a specific resource of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_dataset_resource(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{integer, the index of the resource to access}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Resource object, the selected resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_dataset_resources"></a>}}
\if{latex}{\out{\hypertarget{method-list_dataset_resources}{}}}
\subsection{Method \code{list_dataset_resources()}}{
Get all resources of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$list_dataset_resources(pattern = NULL, format = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{character, regex pattern in resource name}

\item{\code{format}}{character, format of the resources}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of Resource objects, all resources available in the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-number_of_resources"></a>}}
\if{latex}{\out{\hypertarget{method-number_of_resources}{}}}
\subsection{Method \code{number_of_resources()}}{
Get number of dataset resources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$number_of_resources()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of Resource objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_dataset_resource"></a>}}
\if{latex}{\out{\hypertarget{method-delete_dataset_resource}{}}}
\subsection{Method \code{delete_dataset_resource()}}{
Delete a resource by its index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$delete_dataset_resource(index = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index, }}{the index of the resource to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_all_dataset_resources"></a>}}
\if{latex}{\out{\hypertarget{method-delete_all_dataset_resources}{}}}
\subsection{Method \code{delete_all_dataset_resources()}}{
Delete all resources from a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$delete_all_dataset_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the dataset page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-get_configuration}{}}}
\subsection{Method \code{get_configuration()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_configuration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_dataset_date_range"></a>}}
\if{latex}{\out{\hypertarget{method-get_dataset_date_range}{}}}
\subsection{Method \code{get_dataset_date_range()}}{
Get the dataset date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_dataset_date_range()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a date, the dataset date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_maintainer"></a>}}
\if{latex}{\out{\hypertarget{method-get_maintainer}{}}}
\subsection{Method \code{get_maintainer()}}{
Get the dataset maintainer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_maintainer()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An User object, the maintainer of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_container"></a>}}
\if{latex}{\out{\hypertarget{method-get_container}{}}}
\subsection{Method \code{get_container()}}{
Get the dataset organization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$get_container()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an Container object, the container where the data is
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with dataset field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
