% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{delete_all_ridl_resource.RIDLDataset}
\alias{delete_all_ridl_resource.RIDLDataset}
\alias{delete_all_ridl_resource}
\alias{delete_all_ridl_resource.default}
\title{Delete all resources from dataset}
\usage{
\method{delete_all_ridl_resource}{RIDLDataset}(dataset)

delete_all_ridl_resource(dataset)

\method{delete_all_ridl_resource}{default}(dataset)
}
\arguments{
\item{dataset}{A RIDLDataset, the dataset with resources remove}
}
\value{
a RIDLDataset without resources
}
\description{
Delete all resources from dataset
}
