% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{delete_resources.RIDLDataset}
\alias{delete_resources.RIDLDataset}
\alias{delete_resources}
\alias{delete_resources.default}
\title{Delete all resource from dataset}
\usage{
\method{delete_resources}{RIDLDataset}(dataset)

delete_resources(dataset)

\method{delete_resources}{default}(dataset)
}
\arguments{
\item{dataset}{A Dataset, the dataset to remove}
}
\value{
Dataset without resources
}
\description{
Delete all resource from dataset
}
