% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{delete_ridl_config}
\alias{delete_ridl_config}
\title{Delete the ridl config}
\usage{
delete_ridl_config()
}
\description{
Delete the configuration settings for using RIDL
}
\details{
Delete RIDL config
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
set_ridl_config()
get_ridl_config()

delete_ridl_config()
get_ridl_config()
}
}
