% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{delete_ridl_resource.RIDLDataset}
\alias{delete_ridl_resource.RIDLDataset}
\alias{delete_ridl_resource}
\alias{delete_ridl_resource.default}
\title{Delete resource from a dataset}
\usage{
\method{delete_ridl_resource}{RIDLDataset}(dataset, n)

delete_ridl_resource(dataset, n)

\method{delete_ridl_resource}{default}(dataset, n)
}
\arguments{
\item{dataset}{RIDLDataset the dataset from which we one to remove the resource}

\item{n}{integer, the index of the resource to be removed}
}
\value{
Dataset the dataset without the resource
}
\description{
Delete resource from a dataset
}
