% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{get_container.RIDLDataset}
\alias{get_container.RIDLDataset}
\alias{get_container}
\alias{get_container.default}
\title{Dataset container}
\usage{
\method{get_container}{RIDLDataset}(dataset)

get_container(dataset)

\method{get_container}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}
}
\value{
A \code{RIDLContainer}
}
\description{
Get the container where the data is share
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 res <- search_dataset(rows = 3L, visibility = "public")
 get_container(res[[1]])
}
}
