% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_container_name}
\alias{get_container_name}
\title{Dataset organization name}
\usage{
get_container_name(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character The name of the organization sharing the data
}
\description{
Get the organization sharing the data
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 res <- search_dataset(rows = 3L)
 get_container_name(res[[1]])
}
}
