% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{get_dataset.RIDLResource}
\alias{get_dataset.RIDLResource}
\alias{get_dataset}
\alias{get_dataset.default}
\title{Get the dataset containing the resource}
\usage{
\method{get_dataset}{RIDLResource}(resource)

get_dataset(resource)

\method{get_dataset}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
a RIDLDataset, the dataset containing the resource
}
\description{
Get the dataset containing the resource
}
