% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_container}
\alias{get_dataset_container}
\title{Dataset container}
\usage{
get_dataset_container(dataset)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}
}
\value{
A \code{RIDLContainer}
}
\description{
Get the container where the data is share
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 res <- search_dataset(rows = 3L, visibility = "public")
 get_dataset_container(res[[1]])
}
}
