% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_resource}
\alias{get_dataset_resource}
\title{Add resource to dataset}
\usage{
get_dataset_resource(dataset, index)

get_dataset_resource(dataset, index)
}
\arguments{
\item{dataset}{Dataset}

\item{index}{integer, resource position in the dataset}
}
\value{
Resource

Resource
}
\description{
Add resource to dataset

Add resource to dataset
}
