% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{get_dataset_tags_names}
\alias{get_dataset_tags_names}
\title{Dataset tags name}
\usage{
get_dataset_tags_names(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character Tags of the dataset
}
\description{
Gets dataset tags name
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 set_ridl_config()
 res <- search_dataset(rows = 3L)
 get_tags_names(res[[1]])
}
}
