% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{get_file_to_upload.RIDLResource}
\alias{get_file_to_upload.RIDLResource}
\alias{get_file_to_upload}
\alias{get_file_to_upload.default}
\title{Get the file to upload}
\usage{
\method{get_file_to_upload}{RIDLResource}(resource)

get_file_to_upload(resource)

\method{get_file_to_upload}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
A character, the path to file to upload or NULL if not available
}
\description{
Get the file to upload
}
