% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{get_format.RIDLResource}
\alias{get_format.RIDLResource}
\alias{get_format}
\alias{get_format.default}
\title{Get the file format of the resource}
\usage{
\method{get_format}{RIDLResource}(resource)

get_format(resource)

\method{get_format}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
A character, the format of the resource
}
\description{
Get the file format of the resource
}
