% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{get_nth_resource.RIDLDataset}
\alias{get_nth_resource.RIDLDataset}
\alias{get_nth_resource}
\alias{get_nth_resource.default}
\title{Access the nth resource of a dataset or a list of resources}
\usage{
\method{get_nth_resource}{RIDLDataset}(dataset, n)

get_nth_resource(dataset, n)

\method{get_nth_resource}{default}(dataset, n)
}
\arguments{
\item{dataset}{a RIDLDataset or a ridl_resource_list}

\item{n}{integer, resource position in the dataset}
}
\value{
a RIDLResource
}
\description{
Access the nth resource of a dataset or a list of resources
}
