% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{get_ridl_container.RIDLDataset}
\alias{get_ridl_container.RIDLDataset}
\alias{get_ridl_container}
\alias{get_ridl_container.default}
\title{Get the dataset container}
\usage{
\method{get_ridl_container}{RIDLDataset}(dataset)

get_ridl_container(dataset)

\method{get_ridl_container}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}
}
\value{
A \code{RIDLContainer}
}
\description{
Get the container where the data is share
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 res <- search_ridl_dataset(rows = 3L, visibility = "public")
 get_ridl_container(res[[1]])
}
}
