% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{get_ridl_resource.RIDLDataset}
\alias{get_ridl_resource.RIDLDataset}
\alias{get_ridl_resource}
\alias{get_ridl_resource.default}
\title{Access the nth resource of a dataset}
\usage{
\method{get_ridl_resource}{RIDLDataset}(dataset, n)

get_ridl_resource(dataset, n)

\method{get_ridl_resource}{default}(dataset, n)
}
\arguments{
\item{dataset}{a RIDLDataset}

\item{n}{integer, resource position in the dataset}
}
\value{
a RIDLResource
}
\description{
Access the nth resource of a dataset
}
