% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{get_sheets.RIDLResource}
\alias{get_sheets.RIDLResource}
\alias{get_sheets}
\alias{get_sheets.default}
\title{Get the names of the sheets of XLS(X) resources}
\usage{
\method{get_sheets}{RIDLResource}(resource, format = NULL, download_folder = NULL, quiet = TRUE)

get_sheets(resource, format, download_folder, quiet)

\method{get_sheets}{default}(resource, format, download_folder, quiet)
}
\arguments{
\item{resource}{Resource, a RIDL resource}

\item{format}{character, specify file format in case the automatic reader doesn't work as expected}

\item{download_folder}{character, path of the directory where you will store the data}

\item{quiet}{logical, no progress bar from download (default = FALSE)}
}
\value{
the names of the sheets of XLS(X) resources
}
\description{
Get the names of the sheets of XLS(X) resources
}
