% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keywords.R
\docType{data}
\name{keywords}
\alias{keywords}
\title{Topic classifications}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 57 rows and 2 columns.
}
\usage{
keywords
}
\description{
A tibble mapping metadata keyword labels to their internal codes on the system.
}
\keyword{datasets}
