% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{list_dataset_resources}
\alias{list_dataset_resources}
\title{Add resource to dataset}
\usage{
list_dataset_resources(dataset, pattern = NULL, format = NULL)
}
\arguments{
\item{dataset}{Dataset}

\item{pattern}{character, regex pattern in resource name}

\item{format}{character, format of the resources}
}
\value{
resource_list
}
\description{
Add resource to dataset
}
