% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/ridlobject.R
\name{list_datasets.RIDLContainer}
\alias{list_datasets.RIDLContainer}
\alias{list_datasets.default}
\alias{list_datasets}
\title{List RIDL datasets}
\usage{
\method{list_datasets}{RIDLContainer}(container = NULL, configuration = NULL)

\method{list_datasets}{default}(container = NULL, configuration = NULL)

list_datasets(container, configuration)
}
\arguments{
\item{container}{RIDLContainer, the container containing the datasets}

\item{configuration}{a RIDLConfig, the configuration object}
}
\value{
A vector of datasets names
}
\description{
List RIDL datasets
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 list_datasets()
}

}
