% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{list_resources.RIDLDataset}
\alias{list_resources.RIDLDataset}
\alias{list_resources}
\alias{list_resources.default}
\title{List all the resources in a dataset}
\usage{
\method{list_resources}{RIDLDataset}(dataset, pattern = NULL, format = NULL)

list_resources(dataset, pattern, format)

\method{list_resources}{default}(dataset, pattern, format)
}
\arguments{
\item{dataset}{RIDLDataset}

\item{pattern}{character, regex pattern to match resource name}

\item{format}{character, format of the resources}
}
\value{
A ridl_resources_list
}
\description{
List all the resources in a dataset
}
