% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/ridlobject.R
\name{list_ridl_container.default}
\alias{list_ridl_container.default}
\alias{list_ridl_container}
\title{List RIDL containers}
\usage{
\method{list_ridl_container}{default}(
  sort = c("title asc", "name", "package_count", "title"),
  configuration = NULL
)

list_ridl_container(sort, configuration)
}
\arguments{
\item{sort}{character how to sort the results. Default is "name asc"}

\item{configuration}{RIDLConfig, a configuration}
}
\value{
A list of containers on RIDL
}
\description{
List RIDL containers
}
