% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/ridlobject.R
\name{list_ridl_dataset.RIDLContainer}
\alias{list_ridl_dataset.RIDLContainer}
\alias{list_ridl_dataset.default}
\alias{list_ridl_dataset}
\title{List RIDL datasets}
\usage{
\method{list_ridl_dataset}{RIDLContainer}(container = NULL, configuration = NULL)

\method{list_ridl_dataset}{default}(container = NULL, configuration = NULL)

list_ridl_dataset(container, configuration)
}
\arguments{
\item{container}{RIDLContainer, the container containing the datasets}

\item{configuration}{a RIDLConfig, the configuration object}
}
\value{
A vector of datasets names
}
\description{
List RIDL datasets
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 list_ridl_dataset()
}

}
