% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{list_ridl_resource.RIDLDataset}
\alias{list_ridl_resource.RIDLDataset}
\alias{list_ridl_resource}
\alias{list_ridl_resource.default}
\title{List all the resources in a dataset}
\usage{
\method{list_ridl_resource}{RIDLDataset}(dataset, pattern = NULL, format = NULL)

list_ridl_resource(dataset, pattern, format)

\method{list_ridl_resource}{default}(dataset, pattern, format)
}
\arguments{
\item{dataset}{RIDLDataset}

\item{pattern}{character, regex pattern to match resource name}

\item{format}{character, format of the resources}
}
\value{
A ridl_resource_list
}
\description{
List all the resources in a dataset
}
