% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{n_ridl_resource.RIDLDataset}
\alias{n_ridl_resource.RIDLDataset}
\alias{n_ridl_resource}
\alias{n_ridl_resource.default}
\title{Get the number of resources in the dataset}
\usage{
\method{n_ridl_resource}{RIDLDataset}(dataset)

n_ridl_resource(dataset)

\method{n_ridl_resource}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset}
}
\value{
integer, the number of resources
}
\description{
Get the number of resources in the dataset
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 res <- search_ridl_dataset(rows = 3L, visibility = "public")
 n_ridl_resource(res[[2]])
}
}
