% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{pull_container}
\alias{pull_container}
\title{Read a RIDL container}
\usage{
pull_container(
  identifier = NULL,
  include_datasets = FALSE,
  configuration = NULL,
  ...
)
}
\arguments{
\item{identifier}{character resource uuid}

\item{include_datasets}{Logical, include datasets if TRUE}

\item{configuration}{a RIDL configuration object}

\item{...}{Extra parameters}
}
\value{
RIDL container
}
\description{
Read a RIDL container
}
