% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{pull_dataset}
\alias{pull_dataset}
\title{Pull a RIDL dataset}
\usage{
pull_dataset(identifier, configuration = NULL)
}
\arguments{
\item{identifier}{character, name or id of the dataset}

\item{configuration}{RIDLConfig, the configuration to use if any}
}
\value{
Dataset the dataset
}
\description{
Read a RIDL dataset from its name or id
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 res <- pull_dataset("unhcr-mrt-2017-sea-1-1")
 res
}
}
