% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{pull_resource}
\alias{pull_resource}
\title{Pull a RIDL resource}
\usage{
pull_resource(identifier, configuration = NULL)
}
\arguments{
\item{identifier}{character, a RIDLResource id}

\item{configuration}{RIDLConfig, the configuration used}
}
\value{
RIDLResource
}
\description{
Pull a RIDL resource
}
\examples{
\dontrun{
 # Setting the config to use RIDL default server
 set_ridl_config()
 res <- pull_resource("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
 res
}
}
