% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rc_metadata_country}
\alias{rc_metadata_country}
\alias{rc_metadata_sectoral_area}
\title{Visibility metadata}
\format{
A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}
}
\usage{
rc_metadata_country

rc_metadata_sectoral_area
}
\description{
Internal access level

Internal access level
}
\examples{
rc_metadata_country
rc_metadata_sectoral_area
}
\keyword{datasets}
