% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rd_metadata_archived}
\alias{rd_metadata_archived}
\alias{rd_metadata_data_collection_technique}
\alias{rd_metadata_data_sensitivity}
\alias{rd_metadata_external_access_level}
\alias{rd_metadata_identifiability}
\alias{rd_metadata_keywords}
\alias{rd_metadata_operational_purpose_of_data}
\alias{rd_metadata_process_status}
\alias{rd_metadata_sampling_procedure}
\alias{rd_metadata_visibility}
\alias{rd_metadata_uat_archived}
\alias{rd_metadata_uat_data_collection_technique}
\alias{rd_metadata_uat_data_sensitivity}
\alias{rd_metadata_uat_external_access_level}
\alias{rd_metadata_uat_identifiability}
\alias{rd_metadata_uat_keywords}
\alias{rd_metadata_uat_operational_purpose_of_data}
\alias{rd_metadata_uat_process_status}
\alias{rd_metadata_uat_sampling_procedure}
\alias{rd_metadata_uat_visibility}
\title{Dataset archived metadata schema on the production server}
\format{
A tibble with 2 rows and 2 variables:
\describe{
\item{archived_code}{Code of archived}
\item{archived_label}{Label of archived}
}

A tibble with 17 rows and 2 variables:
\describe{
\item{data_collection_technique_code}{Code of data collection technique}
\item{data_collection_technique_label}{Label of data collection technique}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{data_sensitivity_code}{Code of data sensitivity}
\item{data_sensitivity_label}{Label of data sensitivity}
}

A tibble with 6 rows and 2 variables:
\describe{
\item{external_access_level_code}{Name for external access level}
\item{external_access_level_label}{Label for external access level}
}

A tibble with 4 rows and 2 variables:
\describe{
\item{identifiability_code}{Code for identifiability}
\item{identifiability_label}{Label for identifiability}
}

A tibble with 57 rows and 2 variables:
\describe{
\item{keywords_code}{Code for keywords}
\item{keywords_label}{Label for keywords}
}

A tibble with 8 rows and 2 variables:
\describe{
\item{operational_purpose_of_data_code}{Code for operational purpose of data}
\item{operational_purpose_of_data_label}{Label for operational purpose of data}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{process_status_code}{Code of process status}
\item{process_status_label}{Label of process status}
}

A tibble with 17 rows and 2 variables:
\describe{
\item{sampling_procedure_code}{Code for sampling procedure}
\item{sampling_procedure_label}{Label for sampling procedure}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{archived_code}{Code of archived}
\item{archived_label}{Label of archived}
}

A tibble with 17 rows and 2 variables:
\describe{
\item{data_collection_technique_code}{Code of data collection technique}
\item{data_collection_technique_label}{Label of data collection technique}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{data_sensitivity_code}{Code of data sensitivity}
\item{data_sensitivity_label}{Label of data sensitivity}
}

A tibble with 6 rows and 2 variables:
\describe{
\item{external_access_level_code}{Name for external access level}
\item{external_access_level_label}{Label for external access level}
}

A tibble with 4 rows and 2 variables:
\describe{
\item{identifiability_code}{Code for identifiability}
\item{identifiability_label}{Label for identifiability}
}

A tibble with 57 rows and 2 variables:
\describe{
\item{keywords_code}{Code for keywords}
\item{keywords_label}{Label for keywords}
}

A tibble with 8 rows and 2 variables:
\describe{
\item{operational_purpose_of_data_code}{Code for operational purpose of data}
\item{operational_purpose_of_data_label}{Label for operational purpose of data}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{process_status_code}{Code of process status}
\item{process_status_label}{Label of process status}
}

A tibble with 17 rows and 2 variables:
\describe{
\item{sampling_procedure_code}{Code for sampling procedure}
\item{sampling_procedure_label}{Label for sampling procedure}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}
}
\usage{
rd_metadata_archived

rd_metadata_data_collection_technique

rd_metadata_data_sensitivity

rd_metadata_external_access_level

rd_metadata_identifiability

rd_metadata_keywords

rd_metadata_operational_purpose_of_data

rd_metadata_process_status

rd_metadata_sampling_procedure

rd_metadata_visibility

rd_metadata_uat_archived

rd_metadata_uat_data_collection_technique

rd_metadata_uat_data_sensitivity

rd_metadata_uat_external_access_level

rd_metadata_uat_identifiability

rd_metadata_uat_keywords

rd_metadata_uat_operational_purpose_of_data

rd_metadata_uat_process_status

rd_metadata_uat_sampling_procedure

rd_metadata_uat_visibility
}
\description{
Dataset archived metadata schema on the production server

Dataset data collection technique metadata schema on the production server

Dataset data sensitivity metadata schema on the production server

Dataset external access level metadata schema on the production server

Dataset identifiability metadata schema on the production server

Dataset keywords metadata schema on the production server

Dataset operational purpose of data metadata schema on the production server

Dataset process status metadata schema on the production server

Dataset sampling procedure metadata schema on the production server

Dataset visibility metadata schema on the production server

Dataset archived metadata schema on the UAT server

Dataset data collection technique metadata schema on the UAT server

Dataset data sensitivity metadata schema on the UAT server

Dataset external access level metadata schema on the UAT server

Dataset identifiability metadata schema on the UAT server

Dataset keywords metadata schema on the UAT server

Dataset operational purpose of data metadata schema on the UAT server

Dataset process status metadata schema on the UAT server

Dataset sampling procedure metadata schema on the UAT server

Dataset visibility metadata schema on the UAT server
}
\examples{
rd_metadata_archived
rd_metadata_data_collection_technique
rd_metadata_data_sensitivity
rd_metadata_external_access_level
rd_metadata_identifiability
rd_metadata_keywords
rd_metadata_operational_purpose_of_data
rd_metadata_process_status
rd_metadata_sampling_procedure
rd_metadata_visibility
rd_metadata_uat_archived
rd_metadata_uat_data_collection_technique
rd_metadata_uat_data_sensitivity
rd_metadata_uat_external_access_level
rd_metadata_uat_identifiability
rd_metadata_uat_keywords
rd_metadata_uat_operational_purpose_of_data
rd_metadata_uat_process_status
rd_metadata_uat_sampling_procedure
rd_metadata_uat_visibility
}
\keyword{datasets}
