% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{read_resource}
\alias{read_resource}
\title{Read a RIDL Resource}
\usage{
read_resource(
  resource,
  sheet = NULL,
  format = NULL,
  download_folder = NULL,
  force_download = FALSE,
  quiet_download = TRUE,
  ...
)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{sheet}{character, the name of the sheet to read if XLS(X) resources. The first sheet is read by default.}

\item{format}{character, file format, csv, zipped csv, excel, xlsx, zipped shapefile, etc.}

\item{download_folder}{character, the path of the folder to store the
downloaded data}

\item{force_download}{logical, force download if TRUE}

\item{quiet_download}{logical, silent download}

\item{...}{extra parameters}
}
\value{
A \code{tibble}
}
\description{
Read a RIDL Resource
}
