% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridl.R
\name{ridl}
\alias{ridl}
\title{Helper function to make API calls}
\usage{
ridl(action, ...)
}
\arguments{
\item{action}{Operation to execute. See \href{https://docs.ckan.org/en/2.9/api/}{CKAN's API documentation} for details.}

\item{...}{}
}
\value{
httr::response object with the result of the call.
}
\description{

}
