% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/resource.R,
%   R/ridlobject.R
\name{ridl_browse.RIDLContainer}
\alias{ridl_browse.RIDLContainer}
\alias{ridl_browse.RIDLDataset}
\alias{ridl_browse.RIDLResource}
\alias{ridl_browse}
\alias{ridl_browse.default}
\title{Browse a RIDL object}
\usage{
\method{ridl_browse}{RIDLContainer}(x, ...)

\method{ridl_browse}{RIDLDataset}(x, ...)

\method{ridl_browse}{RIDLResource}(x, ...)

ridl_browse(x, ...)

\method{ridl_browse}{default}(x, ...)
}
\arguments{
\item{x}{an RIDL object}

\item{...}{Extra parameters}
}
\value{
Character Tags of the dataset
}
\description{
Browse a RIDL object
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 ridl_config_set()
 res <- ridl_dataset_search(rows = 3L)
 ridl_browse(res[[1]])
}
}
