% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/resource.R,
%   R/ridlobject.R
\name{ridl_clone.RIDLContainer}
\alias{ridl_clone.RIDLContainer}
\alias{ridl_clone.RIDLDataset}
\alias{ridl_clone.RIDLResource}
\alias{ridl_clone}
\alias{ridl_clone.default}
\title{Copy metadata from a RIDL object}
\usage{
\method{ridl_clone}{RIDLContainer}(x, configuration = NULL)

\method{ridl_clone}{RIDLDataset}(x, configuration = NULL)

\method{ridl_clone}{RIDLResource}(x, configuration = NULL)

ridl_clone(x, configuration)

\method{ridl_clone}{default}(x, configuration)
}
\arguments{
\item{x}{an RIDL object}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
Character Tags of the dataset
}
\description{
Copy metadata from a RIDL object
}
\examples{
\dontrun{

 # Setting the config to use RIDL default server
 ridl_config_set()
 res <- ridl_dataset_search(rows = 3L)
 ridl_clone(res[[1]])
}
}
