% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ridl_config}
\alias{ridl_config}
\alias{set_ridl_config}
\alias{get_ridl_config}
\title{Create a RIDL configuration object}
\usage{
ridl_config(ridl_key = NULL, user_agent = NULL)

set_ridl_config(ridl_key = NULL, user_agent = NULL, configuration = NULL)

get_ridl_config()
}
\arguments{
\item{ridl_key}{character, the CKAN API key, it is required to push data into RIDL}

\item{user_agent}{a character, A user agent string}

\item{configuration}{Configuration object.}
}
\value{
A RIDLConfig object

Invisibly returns the ridl config object
}
\description{
Create and RIDL configuration object

Sets the configuration settings for using RIDL.
}
\details{
Setting up a configuration will help you access from a RIDL server
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
set_ridl_config(ridl_key = "xxxxxxxxxx")

# You can check your configuration using \code{get_ridl_config}
config <- get_ridl_config()
config
}
}
