% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ridl_config_delete}
\alias{ridl_config_delete}
\title{Delete the ridl config}
\usage{
ridl_config_delete()
}
\description{
Delete the configuration settings for using RIDL
}
\details{
Delete RIDL config
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
ridl_config_set()
ridl_config_get()

ridl_config_delete()
ridl_config_get()
}
}
