% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container}
\alias{ridl_container}
\title{Create a RIDL container from list}
\usage{
ridl_container(data, configuration = NULL)
}
\arguments{
\item{data}{List, list of data}

\item{configuration}{RIDLConfig, RIDL configuration used}
}
\value{
RIDLContainer the container
}
\description{
Create a RIDL container from list with required fields
}
\examples{
\dontrun{

 dsdata <- list(name = "hum-container",
                title = "Humanitarian container")
 res <- ridl_container(dsdata)
 res
}
}
