% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_copy_metadata}
\alias{ridl_container_copy_metadata}
\title{Copy a container metadata}
\usage{
ridl_container_copy_metadata(container, configuration = NULL)
}
\arguments{
\item{container}{RIDLContainer, the container to copy}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
a RIDLContainer
}
\description{
Copy a container metadata
}
