% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ridl_container_country}
\alias{ridl_container_country}
\title{A dictionnary with the labels and values for container country}
\usage{
ridl_container_country()
}
\value{
a data.frame with the keywords labels and values
}
\description{
A dictionnary with the labels and values for container country
}
