% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_create}
\alias{ridl_container_create}
\alias{rc_create}
\title{Create a container on RIDL}
\usage{
ridl_container_create(container, configuration = NULL)

rc_create(container, configuration = NULL)
}
\arguments{
\item{container}{RIDLContainer, the container}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A HttpResponse object
}
\description{
Create a container on RIDL
}
