% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_exist}
\alias{ridl_container_exist}
\title{Check if a container name or id is available on RIDL}
\usage{
ridl_container_exist(container_name, configuration = NULL)
}
\arguments{
\item{container_name}{character, the name or id of the RIDLContainer}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A logical value, TRUE if the RIDLContainer exists
}
\description{
Check if a container name or id is available on RIDL
}
