% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/ridlobject.R
\name{ridl_container_list.default}
\alias{ridl_container_list.default}
\alias{ridl_container_list}
\title{List RIDL containers}
\usage{
\method{ridl_container_list}{default}(
  sort = c("title asc", "name", "package_count", "title"),
  configuration = NULL
)

ridl_container_list(sort, configuration)
}
\arguments{
\item{sort}{character how to sort the results. Default is "name asc"}

\item{configuration}{RIDLConfig, a configuration}
}
\value{
A list of containers on RIDL
}
\description{
List RIDL containers
}
