% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_patch}
\alias{ridl_container_patch}
\title{Patch a container on RIDL}
\usage{
ridl_container_patch(container, configuration = NULL)
}
\arguments{
\item{container}{RIDLContainer, the container}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A HttpResponse object
}
\description{
Patch a container on RIDL
}
