% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ridl_container_sector}
\alias{ridl_container_sector}
\title{A dictionnary with the labels and values for container sectoral areas}
\usage{
ridl_container_sector()
}
\value{
a data.frame with the keywords labels and values
}
\description{
A dictionnary with the labels and values for container sectoral areas
}
