% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_show}
\alias{ridl_container_show}
\alias{rc_show}
\title{Pull a RIDL container}
\usage{
ridl_container_show(
  identifier = NULL,
  include_datasets = TRUE,
  configuration = NULL,
  ...
)

rc_show(identifier = NULL, include_datasets = TRUE, configuration = NULL, ...)
}
\arguments{
\item{identifier}{character, the container name or id}

\item{include_datasets}{logical, include datasets if TRUE}

\item{configuration}{RIDLConfig, a configuration object}

\item{...}{extra parameters for \code{organization_show} CKAN API endpoint}
}
\value{
A RIDLContainer
}
\description{
Pull a RIDL container
}
