% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset}
\alias{ridl_dataset}
\title{Create a RIDL dataset from list}
\usage{
ridl_dataset(data, configuration = NULL)
}
\arguments{
\item{data}{List, list of data}

\item{configuration}{RIDLConfig, RIDL configuration used}
}
\value{
RIDLDataset the dataset
}
\description{
Create a RIDL dataset from list with required fields
}
\examples{
\dontrun{

 dsdata <- list(name = "hum-dataset",
                title = "Humanitarian dataset")
 res <- ridl_dataset(dsdata)
 res
}
}
