% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_dataset_collaborator_list.RIDLDataset}
\alias{ridl_dataset_collaborator_list.RIDLDataset}
\alias{ridl_dataset_collaborator_list.default}
\title{Get all collaborators for a given dataset}
\usage{
ridl_dataset_collaborator_list.RIDLDataset(
  dataset,
  user = NULL,
  capacity = "member",
  configuration = NULL
)

ridl_dataset_collaborator_list.default(user, capacity, configuration)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{user}{character, id or name of the user}

\item{capacity}{character, the user capacity}
}
\description{
Get all collaborators for a given dataset
}
