% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_dataset_container_get.RIDLDataset}
\alias{ridl_dataset_container_get.RIDLDataset}
\alias{rd_container_get.RIDLDataset}
\alias{ridl_dataset_container_get}
\alias{ridl_dataset_container_get.default}
\alias{rd_container_get}
\alias{rd_container_get.default}
\title{Get the dataset container}
\usage{
\method{ridl_dataset_container_get}{RIDLDataset}(dataset)

\method{rd_container_get}{RIDLDataset}(dataset)

ridl_dataset_container_get(dataset)

\method{ridl_dataset_container_get}{default}(dataset)

rd_container_get(dataset)

\method{rd_container_get}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}
}
\value{
A \code{RIDLContainer}
}
\description{
Get the container where the data is share
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 res <- ridl_dataset_search(rows = 3L, visibility = "public")
 ridl_container_get(res[[1]])
}
}
