% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_dataset_container_set.RIDLDataset}
\alias{ridl_dataset_container_set.RIDLDataset}
\alias{rd_container_set.RIDLDataset}
\alias{ridl_dataset_container_set}
\alias{ridl_dataset_container_set.default}
\alias{rd_container_set}
\alias{rd_container_set.default}
\title{Add a dataset to a container}
\usage{
\method{ridl_dataset_container_set}{RIDLDataset}(dataset, container_name, configuration = NULL)

\method{rd_container_set}{RIDLDataset}(dataset, container_name, configuration = NULL)

ridl_dataset_container_set(dataset, container_name, configuration)

\method{ridl_dataset_container_set}{default}(dataset, container_name, configuration)

rd_container_set(dataset, container_name, configuration)

\method{rd_container_set}{default}(dataset, container_name, configuration)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{container_name}{charater, A valid RIDL container name, you can use
\code{ridl_container_list()} to have the list of all containers}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A RIDLDataset
}
\description{
Add a dataset to a container
}
\examples{
\dontrun{
 ds <- ridl_dataset(list(name = "cool-dataset"))
 ridl_container_set(ds, "zimbabwe-shelter-nfi")
}
}
