% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_copy_metadata}
\alias{ridl_dataset_copy_metadata}
\title{Copy a dataset}
\usage{
ridl_dataset_copy_metadata(dataset, configuration = NULL)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset to upload}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
a RIDLDataset
}
\description{
Copy a dataset
}
