% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_exist}
\alias{ridl_dataset_exist}
\alias{rd_exist}
\title{Check if a dataset name or id is available on RIDL}
\usage{
ridl_dataset_exist(dataset_name, configuration = NULL)

rd_exist(dataset_name, configuration = NULL)
}
\arguments{
\item{dataset_name}{character, the name or id of the RIDLDataset}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A logical value, TRUE if the RIDLDataset exists
}
\description{
Check if a dataset name or id is available on RIDL
}
