% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_dataset_get.RIDLResource}
\alias{ridl_dataset_get.RIDLResource}
\alias{ridl_dataset_get}
\alias{ridl_dataset_get.default}
\title{Get the dataset containing the resource}
\usage{
\method{ridl_dataset_get}{RIDLResource}(resource)

ridl_dataset_get(resource)

\method{ridl_dataset_get}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
a RIDLDataset, the dataset containing the resource
}
\description{
Get the dataset containing the resource
}
