% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ridl_dataset_keywords}
\alias{ridl_dataset_keywords}
\title{A dictionnary with the labels and values for dataset keywords}
\usage{
ridl_dataset_keywords()
}
\value{
a data.frame with the keywords labels and values
}
\description{
A dictionnary with the labels and values for dataset keywords
}
