% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/ridlobject.R
\name{ridl_dataset_list.RIDLContainer}
\alias{ridl_dataset_list.RIDLContainer}
\alias{rd_list.RIDLContainer}
\alias{ridl_dataset_list.default}
\alias{rd_list.default}
\alias{ridl_dataset_list}
\alias{rd_list}
\title{List RIDL datasets}
\usage{
\method{ridl_dataset_list}{RIDLContainer}(container = NULL, user = NULL, configuration = NULL)

\method{rd_list}{RIDLContainer}(container = NULL, user = NULL, configuration = NULL)

\method{ridl_dataset_list}{default}(container = NULL, user = NULL, configuration = NULL)

\method{rd_list}{default}(container = NULL, user = NULL, configuration = NULL)

ridl_dataset_list(container, user, configuration)

rd_list(container, user, configuration)
}
\arguments{
\item{container}{RIDLContainer, the container containing the datasets}

\item{user}{character, List datasets from which the user is a member}

\item{configuration}{a RIDLConfig, the configuration object}
}
\value{
A vector of datasets names
}
\description{
List RIDL datasets
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 ridl_config_set()
 ct <- ridl_container_show("africa")
 ridl_dataset_list(ct)
}

}
