% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_patch}
\alias{ridl_dataset_patch}
\title{Patch a dataset on RIDL}
\usage{
ridl_dataset_patch(dataset, configuration = NULL)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset to upload}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\description{
Patch a dataset on RIDL
}
