% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_dataset_resource_add.RIDLDataset}
\alias{ridl_dataset_resource_add.RIDLDataset}
\alias{rd_resource_add.RIDLDataset}
\alias{ridl_dataset_resource_add}
\alias{ridl_dataset_resource_add.default}
\alias{rd_resource_add}
\alias{rd_resource_add.default}
\title{Add a Resource to a dataset}
\usage{
\method{ridl_dataset_resource_add}{RIDLDataset}(
  dataset,
  resource,
  ignore_dataset_id = FALSE,
  configuration = NULL
)

\method{rd_resource_add}{RIDLDataset}(
  dataset,
  resource,
  ignore_dataset_id = FALSE,
  configuration = NULL
)

ridl_dataset_resource_add(dataset, resource, ignore_dataset_id, configuration)

\method{ridl_dataset_resource_add}{default}(dataset, resource, ignore_dataset_id, configuration)

rd_resource_add(dataset, resource, ignore_dataset_id, configuration)

\method{rd_resource_add}{default}(dataset, resource, ignore_dataset_id, configuration)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{resource}{RIDLResource, the resource}

\item{ignore_dataset_id}{logical, ignore the dataset id}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A RIDLDataset
}
\description{
Add a Resource to a dataset
}
