% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_dataset_resource_count.RIDLDataset}
\alias{ridl_dataset_resource_count.RIDLDataset}
\alias{rd_resource_count.RIDLDataset}
\alias{ridl_dataset_resource_count}
\alias{ridl_dataset_resource_count.default}
\alias{rd_resource_count}
\alias{rd_resource_count.default}
\title{Get the number of resources in the dataset}
\usage{
\method{ridl_dataset_resource_count}{RIDLDataset}(dataset)

\method{rd_resource_count}{RIDLDataset}(dataset)

ridl_dataset_resource_count(dataset)

\method{ridl_dataset_resource_count}{default}(dataset)

rd_resource_count(dataset)

\method{rd_resource_count}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset}
}
\value{
integer, the number of resources
}
\description{
Get the number of resources in the dataset
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 res <- ridl_dataset_search(rows = 3L, visibility = "public")
 ridl_dataset_resource_count(res[[2]])
}
}
