% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_search}
\alias{ridl_dataset_search}
\title{Search for datasets on RIDL}
\usage{
ridl_dataset_search(
  query = "*:*",
  visibility = c("all", "public", "restricted"),
  filter_query = "",
  rows = 10L,
  start = 0L,
  page_size = 1000L,
  configuration = NULL,
  progress = FALSE,
  ...
)
}
\arguments{
\item{query}{character Query terms, use solr format
and default to "\emph{:}" (match everything)}

\item{visibility}{character, either all, public or restricted}

\item{filter_query}{character, filter query results}

\item{rows}{integer, number of matching records to return. Defaults to 10.}

\item{start}{integer, the offset in the complete result for where
the set of returned datasets should begin.}

\item{page_size}{integer, Size of page to return. Defaults to 1000.}

\item{configuration}{RIDLConfig object.}

\item{progress}{logical, progress bar. Default to FALSE}

\item{...}{Extra parameters for \code{package_search} endpoints}
}
\value{
A list of RIDL datasets
}
\description{
Search for datasets on RIDL
}
\details{
Search and find datasets on RIDL
}
\examples{
\dontrun{
 # Setting the config to use RIDL default server
 search_datasets("displaced nigeria", rows = 3L)
}

}
