% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_show}
\alias{ridl_dataset_show}
\alias{rd_show}
\title{Pull and show a RIDL dataset}
\usage{
ridl_dataset_show(identifier, configuration = NULL)

rd_show(identifier, configuration = NULL)
}
\arguments{
\item{identifier}{character, name or id of the dataset}

\item{configuration}{RIDLConfig, the configuration to use if any}
}
\value{
Dataset the dataset
}
\description{
Read a RIDL dataset from its name or id
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 res <- ridl_dataset_show("unhcr-mrt-2017-sea-1-1")
 res
}
}
