% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{ridl_dataset_update}
\alias{ridl_dataset_update}
\title{Update a dataset on RIDL}
\usage{
ridl_dataset_update(dataset, progress = FALSE, configuration = NULL)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset to upload}

\item{progress}{logical, progress bar. Default to FALSE}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\description{
Update a dataset on RIDL
}
