% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource}
\alias{ridl_resource}
\title{Create a RIDL resource from list}
\usage{
ridl_resource(data, configuration = NULL)
}
\arguments{
\item{data}{list, list of data}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
Resource the resource
}
\description{
Create a RIDL resource from list with required fields
}
\examples{
\dontrun{

 rsdata <- list(name = "hum-resource",
                title = "Humanitarian resource")
 res <- ridl_resource(rsdata)
 res
}
}
