% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_copy_metadata}
\alias{ridl_resource_copy_metadata}
\title{Copy a resource metadata}
\usage{
ridl_resource_copy_metadata(resource, configuration = NULL)
}
\arguments{
\item{resource}{RIDLResource, the resource to upload}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
a RIDLResource
}
\description{
Copy a resource metadata
}
