% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_create}
\alias{ridl_resource_create}
\alias{rr_create}
\title{Create a resource on RIDL}
\usage{
ridl_resource_create(resource, dataset, configuration = NULL)

rr_create(resource, dataset, configuration = NULL)
}
\arguments{
\item{resource}{RIDLResource, a resource object}

\item{dataset}{RIDLDataset, the dataset where you can share the dataset}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
RIDLResource, the resource
}
\description{
Create a resource on RIDL
}
