% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_dataset.RIDLResource}
\alias{ridl_resource_dataset.RIDLResource}
\alias{rr_dataset}
\alias{rr_dataset.default}
\title{Get the dataset containing the resource}
\usage{
ridl_resource_dataset.RIDLResource(resource)

rr_dataset(resource)

\method{rr_dataset}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
a RIDLDataset, the dataset containing the resource
}
\description{
Get the dataset containing the resource
}
