% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_resource_delete.RIDLDataset}
\alias{ridl_resource_delete.RIDLDataset}
\alias{ridl_resource_delete}
\alias{ridl_resource_delete.default}
\title{Delete resource from a dataset}
\usage{
\method{ridl_resource_delete}{RIDLDataset}(dataset, n)

ridl_resource_delete(dataset, n)

\method{ridl_resource_delete}{default}(dataset, n)
}
\arguments{
\item{dataset}{RIDLDataset the dataset from which we one to remove the resource}

\item{n}{integer, the index of the resource to be removed}
}
\value{
Dataset the dataset without the resource
}
\description{
Delete resource from a dataset
}
