% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_resource_delete_all.RIDLDataset}
\alias{ridl_resource_delete_all.RIDLDataset}
\alias{ridl_resource_delete_all}
\alias{ridl_resource_delete_all.default}
\title{Delete all resources from dataset}
\usage{
\method{ridl_resource_delete_all}{RIDLDataset}(dataset)

ridl_resource_delete_all(dataset)

\method{ridl_resource_delete_all}{default}(dataset)
}
\arguments{
\item{dataset}{A RIDLDataset, the dataset with resources remove}
}
\value{
a RIDLDataset without resources
}
\description{
Delete all resources from dataset
}
