% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_download.RIDLResource}
\alias{ridl_resource_download.RIDLResource}
\alias{rr_download.RIDLResource}
\alias{ridl_resource_download}
\alias{ridl_resource_download.default}
\alias{rr_download}
\alias{rr_download.default}
\title{Download a RIDL resource}
\usage{
\method{ridl_resource_download}{RIDLResource}(
  resource,
  folder = NULL,
  filename = NULL,
  quiet = FALSE,
  force = FALSE,
  ...
)

\method{rr_download}{RIDLResource}(
  resource,
  folder = NULL,
  filename = NULL,
  quiet = FALSE,
  force = FALSE,
  ...
)

ridl_resource_download(resource, folder, filename, quiet, force, ...)

\method{ridl_resource_download}{default}(resource, folder, filename, quiet, force, ...)

rr_download(resource, folder, filename, quiet, force, ...)

\method{rr_download}{default}(resource, folder, filename, quiet, force, ...)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{folder}{character, path of the directory where you will store the data}

\item{filename}{character, name of the file you will download}

\item{quiet}{logical, no progress bar from download (default = \code{FALSE})}

\item{force}{logical, force download (default = \code{FALSE})}

\item{...}{extra paramaters}
}
\value{
character, the download folder path
}
\description{
Download a RIDL resource into a specific folder
}
\examples{
\dontrun{
#Setting the config to use RIDL default server
 res <- ridl_resource_show("98aa1742-b5d3-40c3-94c6-01e31ded6e84")
 ridl_resource_download(res, folder = "/tmp")
}
}
