% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_excel_sheets.RIDLResource}
\alias{ridl_resource_excel_sheets.RIDLResource}
\alias{rr_excel_sheets.RIDLResource}
\alias{ridl_resource_excel_sheets}
\alias{ridl_resource_excel_sheets.default}
\alias{rr_excel_sheets}
\alias{rr_excel_sheets.default}
\title{Get the names of the sheets of XLS(X) resources}
\usage{
\method{ridl_resource_excel_sheets}{RIDLResource}(
  resource,
  format = NULL,
  download_folder = NULL,
  quiet = TRUE
)

\method{rr_excel_sheets}{RIDLResource}(resource, format = NULL, download_folder = NULL, quiet = TRUE)

ridl_resource_excel_sheets(resource, format, download_folder, quiet)

\method{ridl_resource_excel_sheets}{default}(resource, format, download_folder, quiet)

rr_excel_sheets(resource, format, download_folder, quiet)

\method{rr_excel_sheets}{default}(resource, format, download_folder, quiet)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{format}{character, specify file format in case the automatic reader doesn't work as expected}

\item{download_folder}{character, path of the directory where you will store the data}

\item{quiet}{logical, no progress bar from download (default = FALSE)}
}
\value{
the names of the sheets of XLS(X) resources
}
\description{
Get the names of the sheets of XLS(X) resources
}
