% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_exist}
\alias{ridl_resource_exist}
\title{Check if a resource id is available on RIDL}
\usage{
ridl_resource_exist(resource_id, configuration = NULL)
}
\arguments{
\item{resource_id}{character, the id of the RIDLResource}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A logical value, TRUE if the RIDLResource exists
}
\description{
Check if a resource id is available on RIDL
}
