% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_file_format.RIDLResource}
\alias{ridl_resource_file_format.RIDLResource}
\alias{rr_file_format.RIDLResource}
\alias{ridl_resource_file_format}
\alias{rr_file_format}
\alias{ridl_resource_file_format.default}
\alias{rr_file_format.default}
\title{Get the file format of the resource}
\usage{
\method{ridl_resource_file_format}{RIDLResource}(resource)

\method{rr_file_format}{RIDLResource}(resource)

ridl_resource_file_format(resource)

rr_file_format(resource)

\method{ridl_resource_file_format}{default}(resource)

\method{rr_file_format}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}
}
\value{
A character, the format of the resource
}
\description{
Get the file format of the resource
}
