% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_file_to_upload_set.RIDLResource}
\alias{ridl_resource_file_to_upload_set.RIDLResource}
\alias{rr_file_to_upload_set.RIDLResource}
\alias{ridl_resource_file_to_upload<-.RIDLResource}
\alias{rr_file_to_upload<-.RIDLResource}
\alias{ridl_resource_file_to_upload_get.RIDLResource}
\alias{rr_file_to_upload_get.RIDLResource}
\alias{ridl_resource_file_to_upload_set}
\alias{ridl_resource_file_to_upload_set.default}
\alias{rr_file_to_upload_set}
\alias{rr_file_to_upload_set.default}
\alias{ridl_resource_file_to_upload_get}
\alias{ridl_resource_file_to_upload_get.default}
\alias{rr_file_to_upload_get}
\alias{rr_file_to_upload_get.default}
\alias{ridl_resource_file_to_upload}
\alias{ridl_resource_file_to_upload.default}
\alias{ridl_resource_file_to_upload<-}
\alias{ridl_resource_file_to_upload<-.default}
\alias{rr_file_to_upload<-}
\alias{rr_file_to_upload<-.default}
\alias{rr_file_to_upload}
\alias{rr_file_to_upload.default}
\title{Get and set the file to upload}
\usage{
\method{ridl_resource_file_to_upload_set}{RIDLResource}(resource, file_to_upload)

\method{rr_file_to_upload_set}{RIDLResource}(resource, file_to_upload)

\method{ridl_resource_file_to_upload}{RIDLResource}(resource) <- value

\method{rr_file_to_upload}{RIDLResource}(resource) <- value

\method{ridl_resource_file_to_upload_get}{RIDLResource}(resource)

\method{rr_file_to_upload_get}{RIDLResource}(resource)

ridl_resource_file_to_upload_set(resource, file_to_upload)

\method{ridl_resource_file_to_upload_set}{default}(resource, file_to_upload)

rr_file_to_upload_set(resource, file_to_upload)

\method{rr_file_to_upload_set}{default}(resource, file_to_upload)

ridl_resource_file_to_upload_get(resource)

\method{ridl_resource_file_to_upload_get}{default}(resource)

rr_file_to_upload_get(resource)

\method{rr_file_to_upload_get}{default}(resource)

ridl_resource_file_to_upload(resource)

\method{ridl_resource_file_to_upload}{default}(resource)

ridl_resource_file_to_upload(resource) <- value

\method{ridl_resource_file_to_upload}{default}(resource) <- value

rr_file_to_upload(resource) <- value

\method{rr_file_to_upload}{default}(resource) <- value

rr_file_to_upload(resource)

\method{rr_file_to_upload}{default}(resource)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{file_to_upload}{character the path to the file to upload}

\item{value}{character, the path of the file to upload}
}
\value{
A RIDLResource with a file

A character, the path to file to upload or NULL if not available
}
\description{
Get and set the file to upload

Get the file to upload
}
