% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_resource_get.RIDLDataset}
\alias{ridl_resource_get.RIDLDataset}
\alias{ridl_resource_get}
\alias{ridl_resource_get.default}
\title{Access the nth resource of a dataset}
\usage{
\method{ridl_resource_get}{RIDLDataset}(dataset, n)

ridl_resource_get(dataset, n)

\method{ridl_resource_get}{default}(dataset, n)
}
\arguments{
\item{dataset}{a RIDLDataset}

\item{n}{integer, resource position in the dataset}
}
\value{
a RIDLResource
}
\description{
Access the nth resource of a dataset
}
