% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_resource_get_all.RIDLDataset}
\alias{ridl_resource_get_all.RIDLDataset}
\alias{ridl_resource_get_all}
\alias{ridl_resource_get_all.default}
\title{List all the resources in a dataset}
\usage{
\method{ridl_resource_get_all}{RIDLDataset}(dataset, pattern = NULL, format = NULL)

ridl_resource_get_all(dataset, pattern, format)

\method{ridl_resource_get_all}{default}(dataset, pattern, format)
}
\arguments{
\item{dataset}{RIDLDataset}

\item{pattern}{character, regex pattern to match resource name}

\item{format}{character, format of the resources}
}
\value{
A ridl_resource_get_all
}
\description{
List all the resources in a dataset
}
