% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{ridl_resource_n.RIDLDataset}
\alias{ridl_resource_n.RIDLDataset}
\alias{ridl_resource_n}
\alias{ridl_resource_n.default}
\title{Get the number of resources in the dataset}
\usage{
\method{ridl_resource_n}{RIDLDataset}(dataset)

ridl_resource_n(dataset)

\method{ridl_resource_n}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset}
}
\value{
integer, the number of resources
}
\description{
Get the number of resources in the dataset
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 res <- ridl_dataset_search(rows = 3L, visibility = "public")
 ridl_resource_n(res[[2]])
}
}
