% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{ridl_resource_read.RIDLResource}
\alias{ridl_resource_read.RIDLResource}
\alias{rr_read.RIDLResource}
\alias{ridl_resource_read}
\alias{ridl_resource_read.default}
\alias{rr_read}
\alias{rr_read.default}
\title{Read a RIDL Resource}
\usage{
\method{ridl_resource_read}{RIDLResource}(
  resource,
  sheet = NULL,
  format = NULL,
  download_folder = NULL,
  force_download = FALSE,
  quiet_download = TRUE,
  ...
)

\method{rr_read}{RIDLResource}(
  resource,
  sheet = NULL,
  format = NULL,
  download_folder = NULL,
  force_download = FALSE,
  quiet_download = TRUE,
  ...
)

ridl_resource_read(
  resource,
  sheet,
  format,
  download_folder,
  force_download,
  quiet_download,
  ...
)

\method{ridl_resource_read}{default}(
  resource,
  sheet,
  format,
  download_folder,
  force_download,
  quiet_download,
  ...
)

rr_read(
  resource,
  sheet,
  format,
  download_folder,
  force_download,
  quiet_download,
  ...
)

\method{rr_read}{default}(
  resource,
  sheet,
  format,
  download_folder,
  force_download,
  quiet_download,
  ...
)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{sheet}{character, the name of the sheet to read if XLS(X) resources. The first sheet is read by default.}

\item{format}{character, file format, csv, zipped csv, excel, xlsx, zipped shapefile, etc.}

\item{download_folder}{character, the path of the folder to store the
downloaded data}

\item{force_download}{logical, force download if TRUE}

\item{quiet_download}{logical, silent download}

\item{...}{extra parameters}
}
\value{
A \code{tibble}
}
\description{
Read a RIDL Resource
}
