% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{RIDLContainer}
\alias{RIDLContainer}
\title{RIDL Container}
\description{
RIDL Container

RIDL Container
}
\details{
RIDLContainer, it contains all the logic for creating, checking,
and updating resources
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{RIDLContainer}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for the Container fields element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLContainer$new()}}
\item \href{#method-list_datasets}{\code{RIDLContainer$list_datasets()}}
\item \href{#method-browse}{\code{RIDLContainer$browse()}}
\item \href{#method-get_required_fields}{\code{RIDLContainer$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{RIDLContainer$check_required_fields()}}
\item \href{#method-as_list}{\code{RIDLContainer$as_list()}}
\item \href{#method-print}{\code{RIDLContainer$print()}}
\item \href{#method-clone}{\code{RIDLContainer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_config">}\href{../../ridl/html/RIDLObject.html#method-get_config}{\code{ridl::RIDLObject$get_config()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_fields">}\href{../../ridl/html/RIDLObject.html#method-get_fields}{\code{ridl::RIDLObject$get_fields()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a Container object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Container object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_datasets"></a>}}
\if{latex}{\out{\hypertarget{method-list_datasets}{}}}
\subsection{Method \code{list_datasets()}}{
Get the list of datasets within the container
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$list_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of RIDLDataset objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the Container page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get container required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a container
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_fields}{}}}
\subsection{Method \code{check_required_fields()}}{
Check container required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$check_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the container
is not missing a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with container field element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLContainer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
