% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{RIDLDataset}
\alias{RIDLDataset}
\title{RIDL Dataset}
\description{
RIDL Dataset

RIDL Dataset
}
\details{
RIDLDataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{RIDLDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resources}}{list of Resource object within the dataset}

\item{\code{data}}{placeholder for Dataset field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RIDLDataset-new}{\code{RIDLDataset$new()}}
\item \href{#method-RIDLDataset-ridl_resource_get}{\code{RIDLDataset$ridl_resource_get()}}
\item \href{#method-RIDLDataset-ridl_resource_get_all}{\code{RIDLDataset$ridl_resource_get_all()}}
\item \href{#method-RIDLDataset-ridl_resource_count}{\code{RIDLDataset$ridl_resource_count()}}
\item \href{#method-RIDLDataset-ridl_resource_delete}{\code{RIDLDataset$ridl_resource_delete()}}
\item \href{#method-RIDLDataset-ridl_resource_delete_all}{\code{RIDLDataset$ridl_resource_delete_all()}}
\item \href{#method-RIDLDataset-ridl_resource_add}{\code{RIDLDataset$ridl_resource_add()}}
\item \href{#method-RIDLDataset-ridl_delete}{\code{RIDLDataset$ridl_delete()}}
\item \href{#method-RIDLDataset-ridl_container_set}{\code{RIDLDataset$ridl_container_set()}}
\item \href{#method-RIDLDataset-ridl_browse}{\code{RIDLDataset$ridl_browse()}}
\item \href{#method-RIDLDataset-ridl_config_get}{\code{RIDLDataset$ridl_config_get()}}
\item \href{#method-RIDLDataset-get_date_range}{\code{RIDLDataset$get_date_range()}}
\item \href{#method-RIDLDataset-ridl_container_get}{\code{RIDLDataset$ridl_container_get()}}
\item \href{#method-RIDLDataset-ridl_container_get_title}{\code{RIDLDataset$ridl_container_get_title()}}
\item \href{#method-RIDLDataset-collaborator_list}{\code{RIDLDataset$collaborator_list()}}
\item \href{#method-RIDLDataset-get_fields}{\code{RIDLDataset$get_fields()}}
\item \href{#method-RIDLDataset-get_required_fields}{\code{RIDLDataset$get_required_fields()}}
\item \href{#method-RIDLDataset-check_required_fields}{\code{RIDLDataset$check_required_fields()}}
\item \href{#method-RIDLDataset-as_list}{\code{RIDLDataset$as_list()}}
\item \href{#method-RIDLDataset-print}{\code{RIDLDataset$print()}}
\item \href{#method-RIDLDataset-clone}{\code{RIDLDataset$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="browse_on_ridl"><a href='../../ridl/html/RIDLObject.html#method-RIDLObject-browse_on_ridl'><code>ridl::RIDLObject$browse_on_ridl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_config"><a href='../../ridl/html/RIDLObject.html#method-RIDLObject-get_config'><code>ridl::RIDLObject$get_config()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Dataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_get"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_get}{}}}
\subsection{Method \code{ridl_resource_get()}}{
Get the nth resource of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_get(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, the index of the resource to access}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Resource object, the selected resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_get_all"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_get_all}{}}}
\subsection{Method \code{ridl_resource_get_all()}}{
Get all resources of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_get_all(pattern = NULL, format = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{character, regex pattern in resource name}

\item{\code{format}}{character, format of the resources}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of Resource objects, all resources available in the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_count"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_count}{}}}
\subsection{Method \code{ridl_resource_count()}}{
Get number of dataset resources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of RIDLResource objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_delete"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_delete}{}}}
\subsection{Method \code{ridl_resource_delete()}}{
Delete a resource by its index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_delete(index = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index, }}{the index of the resource to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_delete_all"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_delete_all}{}}}
\subsection{Method \code{ridl_resource_delete_all()}}{
Delete all resources from a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_delete_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_resource_add"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_resource_add}{}}}
\subsection{Method \code{ridl_resource_add()}}{
Add a resource to a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_resource_add(resource, ignore_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{RIDLResource, the resource}

\item{\code{ignore_dataset_id}}{logical, ignore the dataset id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_delete"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_delete}{}}}
\subsection{Method \code{ridl_delete()}}{
Delete a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_delete(id, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{character, the id or name of the dataset to delete}

\item{\code{configuration}}{RIDLConfig, the configuration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_container_set"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_container_set}{}}}
\subsection{Method \code{ridl_container_set()}}{
Add a container to a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_container_set(container_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{the name of the container to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_browse"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_browse}{}}}
\subsection{Method \code{ridl_browse()}}{
Browse the dataset page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_config_get"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_config_get}{}}}
\subsection{Method \code{ridl_config_get()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_config_get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-get_date_range"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-get_date_range}{}}}
\subsection{Method \code{get_date_range()}}{
Get the dataset date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_date_range()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a date, the dataset date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_container_get"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_container_get}{}}}
\subsection{Method \code{ridl_container_get()}}{
Get the dataset container name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_container_get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a RIDLContainer, the container where the dataset is shared
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-ridl_container_get_title"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-ridl_container_get_title}{}}}
\subsection{Method \code{ridl_container_get_title()}}{
Get the dataset container title

Get the title of the dataset for pretty printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$ridl_container_get_title()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, the title
Get all collaborators for a given dataset

Get all collaborators for a given dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-collaborator_list"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-collaborator_list}{}}}
\subsection{Method \code{collaborator_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$collaborator_list(
  user = NULL,
  capacity = "member",
  configuration = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user}}{character, id or name of the user}

\item{\code{capacity}}{character, the user capacity}

\item{\code{configuration}}{RIDLConfig, the configuration}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a character, the title
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get dataset fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get dataset required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-check_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-check_required_fields}{}}}
\subsection{Method \code{check_required_fields()}}{
Check dataset required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$check_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the dataset
is not missing a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with dataset field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RIDLDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RIDLDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
