% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridlobject.R
\name{RIDLObject}
\alias{RIDLObject}
\title{RIDLObject abstract class}
\description{
RIDLObject abstract class

RIDLObject abstract class
}
\details{
RIDLObject class containing all logic for accessing,
creating, and updating RIDL objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for RIDLObject field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLObject$new()}}
\item \href{#method-as_list}{\code{RIDLObject$as_list()}}
\item \href{#method-browse_on_ridl}{\code{RIDLObject$browse_on_ridl()}}
\item \href{#method-get_config}{\code{RIDLObject$get_config()}}
\item \href{#method-get_fields}{\code{RIDLObject$get_fields()}}
\item \href{#method-get_required_fields}{\code{RIDLObject$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{RIDLObject$check_required_fields()}}
\item \href{#method-print}{\code{RIDLObject$print()}}
\item \href{#method-clone}{\code{RIDLObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new RIDLObject object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a RIDLObject}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A RIDLObject object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get RIDLObject field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with RIDLObject field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse_on_ridl"></a>}}
\if{latex}{\out{\hypertarget{method-browse_on_ridl}{}}}
\subsection{Method \code{browse_on_ridl()}}{
Browse RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$browse_on_ridl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_config}{}}}
\subsection{Method \code{get_config()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$get_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get dataset fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$get_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get dataset required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_fields}{}}}
\subsection{Method \code{check_required_fields()}}{
Check dataset required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$check_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the dataset
is not missing a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
