% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{RIDLResource}
\alias{RIDLResource}
\title{RIDL Resource}
\description{
RIDL Resource

RIDL Resource
}
\details{
RIDLResource, it contains all the logic for creating, checking,
and updating resources
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{RIDLResource}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for Resource field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLResource$new()}}
\item \href{#method-download}{\code{RIDLResource$download()}}
\item \href{#method-download_folder}{\code{RIDLResource$download_folder()}}
\item \href{#method-read}{\code{RIDLResource$read()}}
\item \href{#method-excel_sheets}{\code{RIDLResource$excel_sheets()}}
\item \href{#method-ridl_dataset}{\code{RIDLResource$ridl_dataset()}}
\item \href{#method-file_format}{\code{RIDLResource$file_format()}}
\item \href{#method-check_required_field}{\code{RIDLResource$check_required_field()}}
\item \href{#method-get_file_to_upload}{\code{RIDLResource$get_file_to_upload()}}
\item \href{#method-set_file_to_upload}{\code{RIDLResource$set_file_to_upload()}}
\item \href{#method-check_url_filetoupload}{\code{RIDLResource$check_url_filetoupload()}}
\item \href{#method-get_fields}{\code{RIDLResource$get_fields()}}
\item \href{#method-get_required_fields}{\code{RIDLResource$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{RIDLResource$check_required_fields()}}
\item \href{#method-as_list}{\code{RIDLResource$as_list()}}
\item \href{#method-ridl_browse}{\code{RIDLResource$ridl_browse()}}
\item \href{#method-print}{\code{RIDLResource$print()}}
\item \href{#method-clone}{\code{RIDLResource$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="browse_on_ridl">}\href{../../ridl/html/RIDLObject.html#method-browse_on_ridl}{\code{ridl::RIDLObject$browse_on_ridl()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_config">}\href{../../ridl/html/RIDLObject.html#method-get_config}{\code{ridl::RIDLObject$get_config()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new RIDLResource object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a resource}

\item{\code{configuration}}{a RIDLConfig object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A RIDLResource object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download"></a>}}
\if{latex}{\out{\hypertarget{method-download}{}}}
\subsection{Method \code{download()}}{
Download a RIDL resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$download(
  folder = NULL,
  filename = NULL,
  quiet = TRUE,
  force = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder}}{a character, folder to save the dataset}

\item{\code{filename}}{a character, filename of the dataset}

\item{\code{quiet}}{a logical value, silent download if TRUE}

\item{\code{force}}{a logical value, force download}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a character, the file path
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-download_folder"></a>}}
\if{latex}{\out{\hypertarget{method-download_folder}{}}}
\subsection{Method \code{download_folder()}}{
Get the download folder for the latest downloaded resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$download_folder()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, folder with the latest downloaded resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read a resource object directly into memory
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$read(
  sheet = NULL,
  format = NULL,
  download_folder = NULL,
  force_download = FALSE,
  quiet_download = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sheet}}{a character value, only for resource in Excel format}

\item{\code{format}}{character, specify file format in case the automatic reader doesn't work as expected}

\item{\code{download_folder}}{a character value, folder to save the downloaded resource}

\item{\code{force_download}}{a logical value, if TRUE force download}

\item{\code{quiet_download}}{a logical value, if TRUE silent download}

\item{\code{...}}{other parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-excel_sheets"></a>}}
\if{latex}{\out{\hypertarget{method-excel_sheets}{}}}
\subsection{Method \code{excel_sheets()}}{
Get the list of sheets name of resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$excel_sheets(
  format = NULL,
  download_folder = NULL,
  quiet_download = TRUE,
  force_download = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{format}}{character, specify file format in case the automatic reader doesn't work as expected}

\item{\code{download_folder}}{a character value, folder to save the downloaded resource}

\item{\code{quiet_download}}{a logical value, if TRUE silent download}

\item{\code{force_download}}{a logical value, if TRUE force download}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the names of the sheets of XLS(X) resources
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ridl_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-ridl_dataset}{}}}
\subsection{Method \code{ridl_dataset()}}{
Get the resource dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$ridl_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a RIDLDataset, the dataset containing the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-file_format"></a>}}
\if{latex}{\out{\hypertarget{method-file_format}{}}}
\subsection{Method \code{file_format()}}{
Get the file format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$file_format()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, the file format of the resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_field"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_field}{}}}
\subsection{Method \code{check_required_field()}}{
Check dataset required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$check_required_field(check_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check_dataset_id}}{logical whether to check or not dataset id}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a logical value, TRUE if the the resource is not missing
a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_to_upload"></a>}}
\if{latex}{\out{\hypertarget{method-get_file_to_upload}{}}}
\subsection{Method \code{get_file_to_upload()}}{
Get the file that will be or has been uploaded if any
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$get_file_to_upload()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the file to upload
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_file_to_upload"></a>}}
\if{latex}{\out{\hypertarget{method-set_file_to_upload}{}}}
\subsection{Method \code{set_file_to_upload()}}{
Set the file that will be uploaded
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$set_file_to_upload(file_to_upload)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_to_upload}}{character the path to the file to upload}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_url_filetoupload"></a>}}
\if{latex}{\out{\hypertarget{method-check_url_filetoupload}{}}}
\subsection{Method \code{check_url_filetoupload()}}{
Check if the resource has an url or a file to upload but not both
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$check_url_filetoupload()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the resource don't mix 'url' and
and 'file_to_upload'
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get resource fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$get_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of fields for a resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get resource required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_fields}{}}}
\subsection{Method \code{check_required_fields()}}{
Check resource required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$check_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the resource
is not missing a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get resource field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with resource field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ridl_browse"></a>}}
\if{latex}{\out{\hypertarget{method-ridl_browse}{}}}
\subsection{Method \code{ridl_browse()}}{
Browse the resource page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$ridl_browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Resource object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLResource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
