% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rr_metadata_file_type}
\alias{rr_metadata_file_type}
\alias{rr_metadata_hxlated}
\alias{rr_metadata_identifiability}
\alias{rr_metadata_process_status}
\alias{rr_metadata_type}
\alias{rr_metadata_visibility}
\alias{rr_metadata_uat_confidentiality_status}
\alias{rr_metadata_uat_file_type}
\alias{rr_metadata_uat_hxlated}
\alias{rr_metadata_uat_identifiability}
\alias{rr_metadata_uat_process_status}
\alias{rr_metadata_uat_type}
\alias{rr_metadata_uat_visibility}
\title{Resource file type metadata  schema on the production server}
\format{
A tibble with 8 rows and 2 variables:
\describe{
\item{file_type_code}{Code for file type}
\item{file_type_label}{Label for file type}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{hxlated_code}{Code for hxlated}
\item{hxlated_label}{Label for hxlated}
}

A tibble with 4 rows and 2 variables:
\describe{
\item{identifiability_code}{Code for identifiability}
\item{identifiability_label}{Label for identifiability}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{process_status_code}{Code for process status}
\item{process_status_label}{Label for process status}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{type_code}{Code for type}
\item{type_label}{Label for type}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 4 rows and 2 variables:
\describe{
\item{metadata_uat_confidentiality_status_code}{Code for confidentiality status}
\item{metadata_uat_confidentiality_status_label}{Label for confidentiality status}
}

A tibble with 8 rows and 2 variables:
\describe{
\item{file_type_code}{Code for file type}
\item{file_type_label}{Label for file type}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{hxlated_code}{Code for hxlated}
\item{hxlated_label}{Label for hxlated}
}

A tibble with 4 rows and 2 variables:
\describe{
\item{identifiability_code}{Code for identifiability}
\item{identifiability_label}{Label for identifiability}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{process_status_code}{Code for process status}
\item{process_status_label}{Label for process status}
}

A tibble with 3 rows and 2 variables:
\describe{
\item{type_code}{Code for type}
\item{type_label}{Label for type}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}
}
\usage{
rr_metadata_file_type

rr_metadata_hxlated

rr_metadata_identifiability

rr_metadata_process_status

rr_metadata_type

rr_metadata_visibility

rr_metadata_uat_confidentiality_status

rr_metadata_uat_file_type

rr_metadata_uat_hxlated

rr_metadata_uat_identifiability

rr_metadata_uat_process_status

rr_metadata_uat_type

rr_metadata_uat_visibility
}
\description{
Resource file type metadata schema on the production server

Resource hxlated metadata schema on the production server

Resource identifiability metadata schema on the production server

Resource process status metadata schema on the production server

Resource type metadata schema on the production server

Resource visibility metadata schema on the production server

Resource metadata confidentiality status schema on the UAT server

Resource file type metadata schema on the UAT server

Resource hxlated metadata schema on the UAT server

Resource identifiability metadata schema on the UAT server

Resource process status metadata schema on the UAT server

Resource type metadata schema on the UAT server

Resource visibility metadata schema on the UAT server
}
\examples{
rr_metadata_file_type
rr_metadata_hxlated
rr_metadata_identifiability
rr_metadata_process_status
rr_metadata_type
rr_metadata_visibility
rr_metadata_uat_confidentiality_status
rr_metadata_uat_file_type
rr_metadata_uat_hxlated
rr_metadata_uat_identifiability
rr_metadata_uat_process_status
rr_metadata_uat_type
rr_metadata_uat_visibility
}
\keyword{datasets}
