% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{search_resources}
\alias{search_resources}
\title{Search for resources}
\usage{
search_resources(query = "*:*", configuration = NULL, ...)
}
\arguments{
\item{query}{character, a query}

\item{configuration}{RIDLConfig, a configuration}

\item{...}{extra params}
}
\value{
A list of RIDLDataset
}
\description{
Search for resources
}
\details{
Search and find datasets on RIDL
}
\examples{
\dontrun{
 # Setting the config to use RIDL default server
 search_resources("format:xlsx")
}
}
