% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{search_ridl_resource}
\alias{search_ridl_resource}
\title{Search for RIDL resources}
\usage{
search_ridl_resource(query = "*:*", configuration = NULL, ...)
}
\arguments{
\item{query}{character, a query}

\item{configuration}{RIDLConfig, a configuration}

\item{...}{extra params}
}
\value{
A list of RIDLDataset
}
\description{
Search for RIDL resources
}
\details{
Search and find datasets on RIDL
}
\examples{
\dontrun{
 # Setting the config to use RIDL default server
 search_ridl_resource("format:xlsx")
}

}
