% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{set_container.RIDLDataset}
\alias{set_container.RIDLDataset}
\alias{set_container}
\alias{set_container.default}
\title{Add a dataset to a container}
\usage{
\method{set_container}{RIDLDataset}(dataset, container_name, configuration = NULL)

set_container(dataset, container_name, configuration)

\method{set_container}{default}(dataset, container_name, configuration)
}
\arguments{
\item{dataset}{RIDLDataset, the dataset}

\item{container_name}{charater, A valid RIDL container name, you can use
\code{list_containers()} to have the list of all containers}

\item{configuration}{RIDLConfig, the RIDL configuration}
}
\value{
A RIDLDataset
}
\description{
Add a dataset to a container
}
\examples{
\dontrun{
 ds <- ridl_dataset(list(name = "cool-dataset"))
 set_container(ds, "zimbabwe-shelter-nfi")
}
}
