% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R, R/ridlobject.R
\name{set_file_to_upload.RIDLResource}
\alias{set_file_to_upload.RIDLResource}
\alias{set_file_to_upload}
\alias{set_file_to_upload.default}
\title{Set the file to upload}
\usage{
\method{set_file_to_upload}{RIDLResource}(resource, file_to_upload)

set_file_to_upload(resource, file_to_upload)

\method{set_file_to_upload}{default}(resource, file_to_upload)
}
\arguments{
\item{resource}{RIDLResource, a RIDL resource}

\item{file_to_upload}{character the path to the file to upload}
}
\value{
A RIDLResource with a file
}
\description{
Set the file to upload
}
