% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{set_ridl_config}
\alias{set_ridl_config}
\alias{get_ridl_config}
\title{Set ridl config}
\usage{
set_ridl_config(ridl_key = NULL, user_agent = NULL, configuration = NULL)

get_ridl_config()
}
\arguments{
\item{ridl_key}{Character for the CKAN API key, it is required to push data into RIDL}

\item{user_agent}{a character value, A user agent
you can push metdata and data to RIDL}

\item{configuration}{Configuration object.}
}
\value{
Invisibly returns the ridl config object
}
\description{
Sets the configuration settings for using ridl.
}
\details{
Setting up a configuration will help you access from a RIDL server
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
set_ridl_config(ridl_site = "demo")

# You can check your configuration using \code{get_ridl_config}
config <- get_ridl_config()
config
}
}
