% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{tag_names.RIDLDataset}
\alias{tag_names.RIDLDataset}
\title{Get the dataset tag names}
\usage{
tag_names.RIDLDataset(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character Tags of the dataset
}
\description{
Get the dataset tags name
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 res <- search_dataset(rows = 3L)
 get_tags_names(res[[1]])
}
}
