% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_dataset_container}
\alias{add_dataset_container}
\title{Add organization to dataset}
\usage{
add_dataset_container(dataset, container)
}
\arguments{
\item{dataset}{RIDLDataset}

\item{Container}{charater, valid RIDL container name}
}
\value{
A RIDLDataset
}
\description{
Add organization to dataset
}
\examples{
\dontrun{
}
}
