% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/resource.R,
%   R/utils.R
\name{browse.RIDLContainer}
\alias{browse.RIDLContainer}
\alias{browse.RIDLDataset}
\alias{browse.RIDLResource}
\alias{browse}
\alias{browse.default}
\title{Browse a RIDL object}
\usage{
\method{browse}{RIDLContainer}(x, ...)

\method{browse}{RIDLDataset}(x, ...)

\method{browse}{RIDLResource}(x, ...)

browse(x, ...)

\method{browse}{default}(x, ...)
}
\arguments{
\item{x}{an RIDL object}

\item{...}{Extra parameters}
}
\value{
Character Tags of the dataset
}
\description{
Browse a RIDL object
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 res <- search_dataset(rows = 3L)
 browse(res[[1]])
}
}
