% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{create_dataset}
\alias{create_dataset}
\title{Create a RIDL dataset from list}
\usage{
create_dataset(initial_data)
}
\arguments{
\item{initial_data}{List, list of data}
}
\value{
Dataset the dataset
}
\description{
Create a RIDL dataset from list with required fields
}
\examples{
\dontrun{

 metadata <- list(name = "hum-dataset",
                  date = "09/25/2018",
                  title = "Humanitarian dataset")
 res <- create_data(metdata)
 res
}
}
