% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{.list_containers}
\alias{.list_containers}
\alias{list_containers}
\alias{list_container_names}
\title{List RIDL container}
\usage{
.list_containers(
  sort = "name asc",
  all_fields = FALSE,
  include_dataset_count = TRUE,
  include_groups = FALSE,
  include_user = FALSE,
  include_tags = FALSE,
  configuration = NULL,
  ...
)

list_containers(
  sort = "name asc",
  all_fields = FALSE,
  include_dataset_count = TRUE,
  include_groups = FALSE,
  include_user = FALSE,
  include_tags = FALSE,
  configuration = NULL,
  ...
)

list_container_names(
  sort = "name asc",
  all_fields = FALSE,
  include_dataset_count = TRUE,
  include_groups = FALSE,
  include_user = FALSE,
  include_tags = FALSE,
  configuration = NULL,
  ...
)
}
\arguments{
\item{sort}{Character how to sort the results. Default is "name asc"}

\item{all_fields}{Logical, include all fields}

\item{include_dataset_count}{Logical include count in the result}

\item{include_groups}{Logical, whether or not to include locations}

\item{include_user}{Logical, whether or not to include user}

\item{include_tags}{Logical whether or not to include tags}

\item{configuration}{Configuration}

\item{...}{extra paramaters}
}
\value{
A list of containers on RIDL
}
\description{
List RIDL container
}
