% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{n_resources.RIDLDataset}
\alias{n_resources.RIDLDataset}
\alias{n_resources}
\alias{n_resources.default}
\title{Dataset tags name}
\usage{
\method{n_resources}{RIDLDataset}(dataset)

n_resources(dataset)

\method{n_resources}{default}(dataset)
}
\arguments{
\item{dataset}{RIDLDataset}
}
\value{
integer, the number of resources
}
\description{
Gets dataset tags name
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 res <- search_dataset(rows = 3L, visibility = "public")
 n_resources(res[[2]])
}
}
