% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{pull_container}
\alias{pull_container}
\title{Pull a RIDL container}
\usage{
pull_container(
  identifier = NULL,
  include_datasets = TRUE,
  configuration = NULL,
  ...
)
}
\arguments{
\item{identifier}{character, the container name or id}

\item{include_datasets}{logical, include datasets if TRUE}

\item{configuration}{RIDLConfig, a configuration object}

\item{...}{extra parameters for \code{organization_show} CKAN API endpoint}
}
\value{
A RIDLContainer
}
\description{
Pull a RIDL container
}
