% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{pull_dataset}
\alias{pull_dataset}
\title{Pull RIDL dataset into R}
\usage{
pull_dataset(identifier, configuration = NULL)
}
\arguments{
\item{identifier}{Character dataset keyword}

\item{configuration}{a Configuration object}
}
\value{
Dataset the dataset
}
\description{
Read a RIDL dataset from its name or id
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 set_ridl_config()
 res <- pull_dataset("mali-3wop")
 res
}
}
