% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rc_metadata_country}
\alias{rc_metadata_country}
\alias{rc_metadata_sectoral_area}
\alias{rc_metadata_uat_country}
\alias{rc_metadata_uat_sectoral_area}
\title{Container country metadata schema on the production server}
\format{
A tibble with 219 rows and 2 variables:
\describe{
\item{country_code}{Code for country}
\item{country_label}{Label for country}
}

A tibble with 57 rows and 2 variables:
\describe{
\item{sectoral_area_code}{Code for sectoral_area}
\item{sectoral_area_label}{Label for sectoral_area}
}

A tibble with 219 rows and 2 variables:
\describe{
\item{country_code}{Code for country}
\item{country_label}{Label for country}
}

A tibble with 57 rows and 2 variables:
\describe{
\item{sectoral_area_code}{Code for sectoral_area}
\item{sectoral_area_label}{Label for sectoral_area}
}
}
\usage{
rc_metadata_country

rc_metadata_sectoral_area

rc_metadata_uat_country

rc_metadata_uat_sectoral_area
}
\description{
Container country metadata schema on the production server

Container sectoral_area metadata scheme on production server

Container country metadata schema on the UAT server

Container sectoral_area metadata scheme on UAT server
}
\examples{
rc_metadata_country
rc_metadata_sectoral_area
rc_metadata_uat_country
rc_metadata_uat_sectoral_area
}
\keyword{datasets}
