% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration.R
\name{ridl_config}
\alias{ridl_config}
\alias{ridl_config_set}
\alias{ridl_config_setup}
\alias{ridl_config_get}
\title{Create a RIDL configuration object}
\usage{
ridl_config(site = "prod", key = NULL, user_agent = NULL)

ridl_config_set(
  site = "prod",
  key = NULL,
  user_agent = NULL,
  configuration = NULL
)

ridl_config_setup(
  site = "prod",
  key = NULL,
  user_agent = NULL,
  configuration = NULL
)

ridl_config_get()
}
\arguments{
\item{site}{character, the RIDL instance, prod (production server), test (testing server)}

\item{key}{character, the CKAN API key, it is required to push data into RIDL}

\item{user_agent}{a character, A user agent string}

\item{configuration}{Configuration object.}
}
\value{
A RIDLConfig object

Invisibly returns the ridl config object
}
\description{
Create and RIDL configuration object

Sets the configuration settings for using RIDL.
}
\details{
Setting up a configuration will help you access from a RIDL server
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
ridl_config_set(ridl_key = "xxxxxxxxxx")

# You can check your configuration using \code{ridl_config_get}
config <- ridl_config_get()
config
}
}
