% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R
\name{ridl_container_patch}
\alias{ridl_container_patch}
\alias{rc_patch}
\title{Patch a container on RIDL}
\usage{
ridl_container_patch(
  container,
  country = NULL,
  visible_external = NULL,
  geographic_area = NULL,
  title = NULL,
  name = NULL,
  description = NULL,
  not_used = NULL,
  sectoral_area = NULL,
  population = NULL,
  tag_string = NULL,
  configuration = NULL
)

rc_patch(
  container,
  country = NULL,
  visible_external = NULL,
  geographic_area = NULL,
  title = NULL,
  name = NULL,
  description = NULL,
  not_used = NULL,
  sectoral_area = NULL,
  population = NULL,
  tag_string = NULL,
  configuration = NULL
)
}
\arguments{
\item{container}{RIDLContainer, the container to patch}

\item{country}{character, a vector of UNHCR country code}

\item{visible_external}{logical, Visible to Partner users}

\item{geographic_area}{character, Geographic area referenced.}

\item{title}{character, the title of the container}

\item{name}{character, the name of the container}

\item{description}{character, the description of the data}

\item{not_used}{character, name of the parent container}

\item{sectoral_area}{character, The sectoral areas covered by the datasets included within. Multiple values are allowed.}

\item{population}{character, Description of the population covered by the container}

\item{tag_string}{character, the tag eg. economy, mental health, government}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
A HttpResponse object
}
\description{
Patch a container on RIDL
}
