% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/container.R, R/dataset.R, R/ridlobject.R
\name{ridl_dataset_list.RIDLContainer}
\alias{ridl_dataset_list.RIDLContainer}
\alias{ridl_dataset_list.default}
\alias{ridl_dataset_list}
\title{List RIDL datasets}
\usage{
\method{ridl_dataset_list}{RIDLContainer}(container = NULL, configuration = NULL)

\method{ridl_dataset_list}{default}(container = NULL, configuration = NULL)

ridl_dataset_list(container, configuration)
}
\arguments{
\item{container}{RIDLContainer, the container containing the datasets}

\item{configuration}{a RIDLConfig, the configuration object}
}
\value{
A vector of datasets names
}
\description{
List RIDL datasets
}
\examples{
\dontrun{
# Setting the config to use RIDL default server
 ridl_config_set()
 list_ridl_dataset()
}

}
