% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_create}
\alias{ridl_resource_create}
\title{Create a resource on RIDL}
\usage{
ridl_resource_create(
  resource,
  dataset_id,
  file_path,
  progress = FALSE,
  quiet = TRUE,
  configuration = NULL
)
}
\arguments{
\item{resource}{RIDLResource, a resource object}

\item{dataset_id}{character, the id or the name of the RIDLDataset}

\item{file_path}{character, the path to the file to upload}

\item{progress}{logical, progress bar. Default to FALSE}

\item{quiet}{logical, quiet. Default to TRUE}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
RIDLResource, the resource
}
\description{
Create a resource on RIDL
}
