% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_patch}
\alias{ridl_resource_patch}
\alias{rr_patch}
\title{Patch a resource on RIDL}
\usage{
ridl_resource_patch(
  resource,
  type = NULL,
  file_type = NULL,
  measurement_unit = NULL,
  methodology = NULL,
  date_range_start = NULL,
  date_range_end = NULL,
  version = NULL,
  process_status = NULL,
  identifiability = NULL,
  file_to_upload = NULL,
  url = NULL,
  name = NULL,
  description = NULL,
  format = NULL,
  hxlated = NULL,
  authors = NULL,
  script_field_name = NULL,
  script_methods = NULL,
  script_instructions = NULL,
  script_dependencies = NULL,
  script_zip_package = NULL,
  script_date = NULL,
  license_name = NULL,
  license_uri = NULL,
  visibility = NULL,
  limitations = NULL,
  dimensions = NULL,
  statistical_concept = NULL,
  time_periods = NULL,
  confidentiality = NULL,
  confidentiality_status = NULL,
  confidentiality_note = NULL,
  script_software = NULL,
  source_code_repo = NULL,
  periodicity = NULL,
  kobo_type = NULL,
  kobo_details = NULL,
  configuration = NULL
)

rr_patch(
  resource,
  type = NULL,
  file_type = NULL,
  measurement_unit = NULL,
  methodology = NULL,
  date_range_start = NULL,
  date_range_end = NULL,
  version = NULL,
  process_status = NULL,
  identifiability = NULL,
  file_to_upload = NULL,
  url = NULL,
  name = NULL,
  description = NULL,
  format = NULL,
  hxlated = NULL,
  authors = NULL,
  script_field_name = NULL,
  script_methods = NULL,
  script_instructions = NULL,
  script_dependencies = NULL,
  script_zip_package = NULL,
  script_date = NULL,
  license_name = NULL,
  license_uri = NULL,
  visibility = NULL,
  limitations = NULL,
  dimensions = NULL,
  statistical_concept = NULL,
  time_periods = NULL,
  confidentiality = NULL,
  confidentiality_status = NULL,
  confidentiality_note = NULL,
  script_software = NULL,
  source_code_repo = NULL,
  periodicity = NULL,
  kobo_type = NULL,
  kobo_details = NULL,
  configuration = NULL
)
}
\arguments{
\item{resource}{RIDLResource, the resource to patch}

\item{type}{character, Resource type(*) - The kind of file you want to upload. Allowed values: \code{data} (Data file), \code{attachment} (Additional attachment).}

\item{file_type}{character, File type(*) - Indicates what is contained in the file. Allowed values: \code{microdata} (Microdata), \code{questionnaire} (Questionnaire), \code{report} (Report), \code{sampling_methodology} (Sampling strategy & methodology Description), \code{infographics} (Infographics & Dashboard), \code{script} (Script), \verb{concept note} (Concept Note), \code{other} (Other).}

\item{measurement_unit}{character, the unit of measurement}

\item{methodology}{character, This element documents methodological details on the production of the series or indicator.}

\item{date_range_start}{Date, Data collection first date(*) - Use yyyy-mm-dd format.}

\item{date_range_end}{Date, Data collection last date(*) - Use yyyy-mm-dd format.}

\item{version}{character, Version(*).}

\item{process_status}{character, File process status(*) - Indicates the processing stage of the data. 'Raw' means that the data has not been cleaned since collection. 'In process' means that it is being cleaned. 'Final' means that the dataset is final and ready for use in analytical products. Allowed valued: \code{raw} (Raw-Uncleaned), \code{cleaned} (Cleaned Only), \code{anonymized} (Cleaned & Anonymized).}

\item{identifiability}{character, Identifiability(*) - Indicates if personally identifiable data is contained in the dataset. Allowed values: \code{personally_identifiable} (Personally identifiable), \code{anonymized_enclave} (Anonymized 1st level: Data Enclave - only removed direct identifiers), \code{anonymized_scientific} (Anonymized 2st level: Scientific Use File (SUF)), \code{anonymized_public} (Anonymized 3st level: Public Use File (PUF)).}

\item{file_to_upload}{character, path of the file to upload}

\item{url}{character, Upload - The file name as it will be recorded in the system.}

\item{name}{character, the name of the resource}

\item{description}{character, Description - Some usefule notes about the data.}

\item{format}{character, File format - eg. CSV, XML, or JSON.}

\item{hxlated}{logical, HXL-ated. Allowed values: \code{False} (No), \code{True} (Yes).}

\item{authors}{character, authors}

\item{script_field_name}{character, name of the script field}

\item{script_methods}{character, statistical/analytic methods included in the script}

\item{script_instructions}{character, instructions for running the script. Information on the sequence in which the scripts must be run is critical.}

\item{script_dependencies}{character, dependencies (packages/libraries) that the script relies on.}

\item{script_zip_package}{character, Script ZIP package}

\item{script_date, }{character, date of the script}

\item{license_name}{character, script license name}

\item{license_uri}{character, script license uri}

\item{visibility}{character, Internal Access Level(*). Allowed values: \code{restricted} (Private), \code{public} (Internally Visible).}

\item{limitations}{character, limitations}

\item{dimensions}{character, dimensions}

\item{statistical_concept}{character, statistical concept used}

\item{time_periods}{character, time periods}

\item{confidentiality}{character, confidentiality}

\item{confidentiality_status}{character, confidentiality status}

\item{confidentiality_note}{character, confidentiality note}

\item{script_software}{character, software used to run the script}

\item{source_code_repo}{character, repository (e.g. GitHub repo) where the script has been published.}

\item{periodicity}{character, periodicity}

\item{kobo_type}{character, type}

\item{kobo_details}{character, details}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
RIDLResource, the resource
}
\description{
Patch a resource on RIDL
}
