% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_update}
\alias{ridl_resource_update}
\title{Update a resource on RIDL}
\usage{
ridl_resource_update(resource, file_path, dataset_id, configuration = NULL)
}
\arguments{
\item{resource}{RIDLResource, a resource object}

\item{file_path}{character, the path to the file to upload}

\item{dataset_id}{character, the id or the name of the RIDLDataset}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
RIDLResource, the resource
}
\description{
Update a resource on RIDL
}
