% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{ridl_resource_update}
\alias{ridl_resource_update}
\alias{rr_update}
\title{Update a resource on RIDL}
\usage{
ridl_resource_update(
  resource,
  type,
  file_type,
  measurement_unit,
  methodology,
  date_range_start,
  date_range_end,
  version,
  process_status,
  identifiability,
  visibility,
  file_to_upload = NULL,
  url = NULL,
  name = NULL,
  title = NULL,
  description = NULL,
  format = NULL,
  hxlated = NULL,
  kobo_type = NULL,
  kobo_details = NULL,
  configuration = NULL
)

rr_update(
  resource,
  type,
  file_type,
  measurement_unit,
  methodology,
  date_range_start,
  date_range_end,
  version,
  process_status,
  identifiability,
  visibility,
  file_to_upload = NULL,
  url = NULL,
  name = NULL,
  title = NULL,
  description = NULL,
  format = NULL,
  hxlated = NULL,
  kobo_type = NULL,
  kobo_details = NULL,
  configuration = NULL
)
}
\arguments{
\item{resource}{RIDLResource, the resource to update}

\item{type}{character, Resource type(*) - The kind of file you want to upload. Allowed values: \code{data} (Data file), \code{attachment} (Additional attachment).}

\item{file_type}{character, File type(*) - Indicates what is contained in the file. Allowed values: \code{microdata} (Microdata), \code{questionnaire} (Questionnaire), \code{report} (Report), \code{sampling_methodology} (Sampling strategy & methodology Description), \code{infographics} (Infographics & Dashboard), \code{script} (Script), \verb{concept note} (Concept Note), \code{other} (Other).}

\item{measurement_unit}{character, the unit of measurement}

\item{methodology}{character, This element documents methodological details on the production of the series or indicator.}

\item{date_range_start}{Date, Data collection first date(*) - Use yyyy-mm-dd format.}

\item{date_range_end}{Date, Data collection last date(*) - Use yyyy-mm-dd format.}

\item{version}{character, Version(*).}

\item{process_status}{character, File process status(*) - Indicates the processing stage of the data. 'Raw' means that the data has not been cleaned since collection. 'In process' means that it is being cleaned. 'Final' means that the dataset is final and ready for use in analytical products. Allowed valued: \code{raw} (Raw-Uncleaned), \code{cleaned} (Cleaned Only), \code{anonymized} (Cleaned & Anonymized).}

\item{identifiability}{character, Identifiability(*) - Indicates if personally identifiable data is contained in the dataset. Allowed values: \code{personally_identifiable} (Personally identifiable), \code{anonymized_enclave} (Anonymized 1st level: Data Enclave - only removed direct identifiers), \code{anonymized_scientific} (Anonymized 2st level: Scientific Use File (SUF)), \code{anonymized_public} (Anonymized 3st level: Public Use File (PUF)).}

\item{visibility}{character, Internal Access Level(*). Allowed values: \code{restricted} (Private), \code{public} (Internally Visible).}

\item{file_to_upload}{character, path of the file to upload}

\item{url}{character, Upload - The file name as it will be recorded in the system.}

\item{name}{character, the name of the resource}

\item{title}{character, title of the resource}

\item{description}{character, Description - Some usefule notes about the data.}

\item{format}{character, File format - eg. CSV, XML, or JSON.}

\item{hxlated}{logical, HXL-ated. Allowed values: \code{False} (No), \code{True} (Yes).}

\item{kobo_type}{character, type}

\item{kobo_details}{character, details}

\item{configuration}{RIDLConfig, the configuration}
}
\value{
RIDLResource, the resource
}
\description{
Update a resource on RIDL
}
