% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{RIDLDataset}
\alias{RIDLDataset}
\title{RIDL Dataset}
\description{
RIDL Dataset

RIDL Dataset
}
\details{
RIDLDataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{RIDLDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resources}}{list of Resource object within the dataset}

\item{\code{data}}{placeholder for Dataset field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLDataset$new()}}
\item \href{#method-get_ridl_resource}{\code{RIDLDataset$get_ridl_resource()}}
\item \href{#method-list_ridl_resource}{\code{RIDLDataset$list_ridl_resource()}}
\item \href{#method-n_ridl_resource}{\code{RIDLDataset$n_ridl_resource()}}
\item \href{#method-delete_ridl_resource}{\code{RIDLDataset$delete_ridl_resource()}}
\item \href{#method-delete_all_ridl_resource}{\code{RIDLDataset$delete_all_ridl_resource()}}
\item \href{#method-add_ridl_resource}{\code{RIDLDataset$add_ridl_resource()}}
\item \href{#method-set_ridl_container}{\code{RIDLDataset$set_ridl_container()}}
\item \href{#method-browse}{\code{RIDLDataset$browse()}}
\item \href{#method-get_ridl_config}{\code{RIDLDataset$get_ridl_config()}}
\item \href{#method-get_date_range}{\code{RIDLDataset$get_date_range()}}
\item \href{#method-get_ridl_container}{\code{RIDLDataset$get_ridl_container()}}
\item \href{#method-get_ridl_container_title}{\code{RIDLDataset$get_ridl_container_title()}}
\item \href{#method-get_fields}{\code{RIDLDataset$get_fields()}}
\item \href{#method-get_required_fields}{\code{RIDLDataset$get_required_fields()}}
\item \href{#method-check_required_fields}{\code{RIDLDataset$check_required_fields()}}
\item \href{#method-as_list}{\code{RIDLDataset$as_list()}}
\item \href{#method-print}{\code{RIDLDataset$print()}}
\item \href{#method-clone}{\code{RIDLDataset$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ridl" data-topic="RIDLObject" data-id="get_config">}\href{../../ridl/html/RIDLObject.html#method-get_config}{\code{ridl::RIDLObject$get_config()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Dataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-get_ridl_resource}{}}}
\subsection{Method \code{get_ridl_resource()}}{
Get the nth resource of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_ridl_resource(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, the index of the resource to access}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Resource object, the selected resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-list_ridl_resource}{}}}
\subsection{Method \code{list_ridl_resource()}}{
Get all resources of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$list_ridl_resource(pattern = NULL, format = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{character, regex pattern in resource name}

\item{\code{format}}{character, format of the resources}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of Resource objects, all resources available in the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-n_ridl_resource}{}}}
\subsection{Method \code{n_ridl_resource()}}{
Get number of dataset resources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$n_ridl_resource()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of RIDLResource objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-delete_ridl_resource}{}}}
\subsection{Method \code{delete_ridl_resource()}}{
Delete a resource by its index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$delete_ridl_resource(index = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index, }}{the index of the resource to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_all_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-delete_all_ridl_resource}{}}}
\subsection{Method \code{delete_all_ridl_resource()}}{
Delete all resources from a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$delete_all_ridl_resource()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_ridl_resource"></a>}}
\if{latex}{\out{\hypertarget{method-add_ridl_resource}{}}}
\subsection{Method \code{add_ridl_resource()}}{
Add a resource to a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$add_ridl_resource(resource, ignore_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{RIDLResource, the resource}

\item{\code{ignore_dataset_id}}{logical, ignore the dataset id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ridl_container"></a>}}
\if{latex}{\out{\hypertarget{method-set_ridl_container}{}}}
\subsection{Method \code{set_ridl_container()}}{
Add a container to a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$set_ridl_container(container_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{the name of the container to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the dataset page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ridl_config"></a>}}
\if{latex}{\out{\hypertarget{method-get_ridl_config}{}}}
\subsection{Method \code{get_ridl_config()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_ridl_config()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_date_range"></a>}}
\if{latex}{\out{\hypertarget{method-get_date_range}{}}}
\subsection{Method \code{get_date_range()}}{
Get the dataset date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_date_range()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a date, the dataset date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ridl_container"></a>}}
\if{latex}{\out{\hypertarget{method-get_ridl_container}{}}}
\subsection{Method \code{get_ridl_container()}}{
Get the dataset container name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_ridl_container()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a RIDLContainer, the container where the dataset is shared
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ridl_container_title"></a>}}
\if{latex}{\out{\hypertarget{method-get_ridl_container_title}{}}}
\subsection{Method \code{get_ridl_container_title()}}{
Get the dataset container title

Get the title of the dataset for pretty printing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_ridl_container_title()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a character, the title
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get dataset fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_required_fields}{}}}
\subsection{Method \code{get_required_fields()}}{
Get dataset required fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of required fields for a dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_required_fields"></a>}}
\if{latex}{\out{\hypertarget{method-check_required_fields}{}}}
\subsection{Method \code{check_required_fields()}}{
Check dataset required field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$check_required_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a logical value, TRUE if the the dataset
is not missing a required field and throws an error otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with dataset field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
