% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{RIDLDataset}
\alias{RIDLDataset}
\title{RIDL Dataset}
\description{
RIDL Dataset

RIDL Dataset
}
\details{
RIDLDataset class containing all logic for accessing,
creating, and updating datasets and associated resources.
}
\section{Super class}{
\code{\link[ridl:RIDLObject]{ridl::RIDLObject}} -> \code{RIDLDataset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{resources}}{list of Resource object within the dataset}

\item{\code{data}}{placeholder for Dataset field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLDataset$new()}}
\item \href{#method-get_nth_resource}{\code{RIDLDataset$get_nth_resource()}}
\item \href{#method-list_resources}{\code{RIDLDataset$list_resources()}}
\item \href{#method-n_resources}{\code{RIDLDataset$n_resources()}}
\item \href{#method-delete_nth_resource}{\code{RIDLDataset$delete_nth_resource()}}
\item \href{#method-delete_resources}{\code{RIDLDataset$delete_resources()}}
\item \href{#method-add_resource}{\code{RIDLDataset$add_resource()}}
\item \href{#method-set_container}{\code{RIDLDataset$set_container()}}
\item \href{#method-browse}{\code{RIDLDataset$browse()}}
\item \href{#method-get_configuration}{\code{RIDLDataset$get_configuration()}}
\item \href{#method-get_date_range}{\code{RIDLDataset$get_date_range()}}
\item \href{#method-get_container}{\code{RIDLDataset$get_container()}}
\item \href{#method-as_list}{\code{RIDLDataset$as_list()}}
\item \href{#method-print}{\code{RIDLDataset$print()}}
\item \href{#method-clone}{\code{RIDLDataset$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a dataset}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Dataset object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_nth_resource"></a>}}
\if{latex}{\out{\hypertarget{method-get_nth_resource}{}}}
\subsection{Method \code{get_nth_resource()}}{
Get the nth resource of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_nth_resource(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{integer, the index of the resource to access}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a Resource object, the selected resource
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list_resources"></a>}}
\if{latex}{\out{\hypertarget{method-list_resources}{}}}
\subsection{Method \code{list_resources()}}{
Get all resources of the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$list_resources(pattern = NULL, format = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pattern}}{character, regex pattern in resource name}

\item{\code{format}}{character, format of the resources}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a list of Resource objects, all resources available in the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-n_resources"></a>}}
\if{latex}{\out{\hypertarget{method-n_resources}{}}}
\subsection{Method \code{n_resources()}}{
Get number of dataset resources
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$n_resources()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of RIDLResource objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_nth_resource"></a>}}
\if{latex}{\out{\hypertarget{method-delete_nth_resource}{}}}
\subsection{Method \code{delete_nth_resource()}}{
Delete a resource by its index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$delete_nth_resource(index = 1L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index, }}{the index of the resource to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_resources"></a>}}
\if{latex}{\out{\hypertarget{method-delete_resources}{}}}
\subsection{Method \code{delete_resources()}}{
Delete all resources from a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$delete_resources()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_resource"></a>}}
\if{latex}{\out{\hypertarget{method-add_resource}{}}}
\subsection{Method \code{add_resource()}}{

\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$add_resource(resource, ignore_dataset_id = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{RIDLResource, the resource}

\item{\code{ignore_dataset_id}}{logical, ignore the dataset id

Add a resource to a dataset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_container"></a>}}
\if{latex}{\out{\hypertarget{method-set_container}{}}}
\subsection{Method \code{set_container()}}{
Add a container to a dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$set_container(container_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{container_name}}{the name of the container to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse the dataset page on RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-get_configuration}{}}}
\subsection{Method \code{get_configuration()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_configuration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_date_range"></a>}}
\if{latex}{\out{\hypertarget{method-get_date_range}{}}}
\subsection{Method \code{get_date_range()}}{
Get the dataset date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_date_range()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a date, the dataset date.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_container"></a>}}
\if{latex}{\out{\hypertarget{method-get_container}{}}}
\subsection{Method \code{get_container()}}{
Get the dataset organization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$get_container()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
an Container object, the container where the data is
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get dataset field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with dataset field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
