% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridlobject.R
\name{RIDLObject}
\alias{RIDLObject}
\title{RIDLObject abstract class}
\description{
RIDLObject abstract class

RIDLObject abstract class
}
\details{
RIDLObject class containing all logic for accessing,
creating, and updating RIDL objects.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{placeholder for RIDLObject field element}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RIDLObject$new()}}
\item \href{#method-as_list}{\code{RIDLObject$as_list()}}
\item \href{#method-browse}{\code{RIDLObject$browse()}}
\item \href{#method-get_configuration}{\code{RIDLObject$get_configuration()}}
\item \href{#method-print}{\code{RIDLObject$print()}}
\item \href{#method-clone}{\code{RIDLObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new RIDLObject object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$new(initial_data = NULL, configuration = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{initial_data}}{list with required field to create a RIDLObject}

\item{\code{configuration}}{a Configuration object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A RIDLObject object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as_list"></a>}}
\if{latex}{\out{\hypertarget{method-as_list}{}}}
\subsection{Method \code{as_list()}}{
Get RIDLObject field into list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$as_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a list with RIDLObject field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-browse"></a>}}
\if{latex}{\out{\hypertarget{method-browse}{}}}
\subsection{Method \code{browse()}}{
Browse RIDL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$browse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_configuration"></a>}}
\if{latex}{\out{\hypertarget{method-get_configuration}{}}}
\subsection{Method \code{get_configuration()}}{
Get the current configuration in use
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$get_configuration()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A configuration object, the configuration in use
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print a Dataset object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RIDLObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
