% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{rr_metadata_confidentiality_status}
\alias{rr_metadata_confidentiality_status}
\alias{rr_metadata_file_type}
\alias{rr_metadata_hxlated}
\alias{rr_metadata_identifiability}
\alias{rr_metadata_process_status}
\alias{rr_metadata_type}
\alias{rr_metadata_visibility}
\title{Metadata confidentiality status}
\format{
A tibble with 2 rows and 2 variables:
\describe{
\item{metadata_confidentiality_status_code}{Code for Metadata confidentiality status}
\item{metadata_confidentiality_status_label}{Label for Metadata confidentiality status}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}

A tibble with 2 rows and 2 variables:
\describe{
\item{visibility_code}{Code for visibility}
\item{visibility_label}{Label for visibility}
}
}
\usage{
rr_metadata_confidentiality_status

rr_metadata_file_type

rr_metadata_hxlated

rr_metadata_identifiability

rr_metadata_process_status

rr_metadata_process_status

rr_metadata_type

rr_metadata_visibility
}
\description{
Metadata confidentiality status

Internal access level

Internal access level

Internal access level

Internal access level

Internal access level

Internal access level

Internal access level
}
\examples{
rr_metadata_confidentiality_status
rr_metadata_file_type
rr_metadata_hxlated
rr_metadata_identifiability
rr_metadata_process_status
rr_metadata_process_status
rr_metadata_type
rr_metadata_visibility
}
\keyword{datasets}
