% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R, R/ridlobject.R
\name{tag_names.RIDLDataset}
\alias{tag_names.RIDLDataset}
\alias{tag_names}
\alias{tag_names.default}
\title{Dataset tags name}
\usage{
\method{tag_names}{RIDLDataset}(dataset)

tag_names(dataset)

\method{tag_names}{default}(dataset)
}
\arguments{
\item{dataset}{Dataset}
}
\value{
Character Tags of the dataset
}
\description{
Gets dataset tags name
}
\examples{
\dontrun{
 # Setting the config to use RIDL
 set_ridl_config()
 res <- search_dataset(rows = 3L)
 get_tags_names(res[[1]])
}
}
